+++ mimikatz/mimikatz.vcxproj
@@ -170,6 +170,7 @@
     <ClCompile Include="modules\kuhl_m_ts.c" />
     <ClCompile Include="modules\kuhl_m_vault.c" />
     <ClCompile Include="modules\kuhl_m_minesweeper.c" />
+    <ClCompile Include="modules\lsadump\kuhl_m_lsadump_dc.c" />
     <ClCompile Include="modules\sekurlsa\crypto\kuhl_m_sekurlsa_nt5.c" />
     <ClCompile Include="modules\sekurlsa\crypto\kuhl_m_sekurlsa_nt6.c" />
     <ClCompile Include="modules\sekurlsa\crypto\kuhl_m_sekurlsa_nt63.c" />
@@ -256,6 +257,7 @@
     <ClInclude Include="modules\kuhl_m_ts.h" />
     <ClInclude Include="modules\kuhl_m_vault.h" />
     <ClInclude Include="modules\kuhl_m_minesweeper.h" />
+    <ClInclude Include="modules\lsadump\kuhl_m_lsadump_dc.h" />
     <ClInclude Include="modules\sekurlsa\crypto\kuhl_m_sekurlsa_nt5.h" />
     <ClInclude Include="modules\sekurlsa\crypto\kuhl_m_sekurlsa_nt6.h" />
     <ClInclude Include="modules\sekurlsa\crypto\kuhl_m_sekurlsa_nt63.h" />
