+++ modules/rpc/kull_m_rpc_ms-drsr_c.c
@@ -3,209 +3,238 @@
 #ifdef _M_X64
 typedef struct _ms2Ddrsr_MIDL_TYPE_FORMAT_STRING {
 	SHORT Pad;
-	UCHAR Format[2071];
+	UCHAR Format[1315];
 } ms2Ddrsr_MIDL_TYPE_FORMAT_STRING;
 
 typedef struct _ms2Ddrsr_MIDL_PROC_FORMAT_STRING {
 	SHORT Pad;
-	UCHAR Format[309];
+	UCHAR Format[695];
 } ms2Ddrsr_MIDL_PROC_FORMAT_STRING;
+static const unsigned short drsuapi_FormatStringOffsetTable[] = {0, 60, 104, 134, 202, 258, 288, 318, 348, 378, 408, 438, 468, 536, 566, 596, 626};
+#elif defined _M_IX86
+typedef struct _ms2Ddrsr_MIDL_TYPE_FORMAT_STRING {
+	SHORT Pad;
+	UCHAR Format[1465];
+} ms2Ddrsr_MIDL_TYPE_FORMAT_STRING;
+
+typedef struct _ms2Ddrsr_MIDL_PROC_FORMAT_STRING {
+	SHORT Pad;
+	UCHAR Format[661];
+} ms2Ddrsr_MIDL_PROC_FORMAT_STRING;
+static const unsigned short drsuapi_FormatStringOffsetTable[] = {0, 58, 100, 128, 194, 248, 276, 304, 332, 360, 388, 416, 444, 510, 538, 566, 594};
+#endif
 
 extern const ms2Ddrsr_MIDL_TYPE_FORMAT_STRING ms2Ddrsr__MIDL_TypeFormatString;
 extern const ms2Ddrsr_MIDL_PROC_FORMAT_STRING ms2Ddrsr__MIDL_ProcFormatString;
+extern const MIDL_SERVER_INFO drsuapi_ServerInfo;
+
+static const RPC_DISPATCH_FUNCTION drsuapi_table[] = {NdrServerCall2, NdrServerCall2, NdrServerCall2, NdrServerCall2, NdrServerCall2, NdrServerCall2, NdrServerCall2, NdrServerCall2, NdrServerCall2, NdrServerCall2, NdrServerCall2, NdrServerCall2, NdrServerCall2, NdrServerCall2, NdrServerCall2, NdrServerCall2, NdrServerCall2, 0};
+static const RPC_DISPATCH_TABLE drsuapi_v4_0_DispatchTable = {17, (RPC_DISPATCH_FUNCTION *) drsuapi_table};
+static const RPC_SERVER_INTERFACE drsuapi___RpcServerInterface = {sizeof(RPC_SERVER_INTERFACE), {{0xe3514235, 0x4b06, 0x11d1, {0xab, 0x04, 0x00, 0xc0, 0x4f, 0xc2, 0xdc, 0xd2}}, {4, 0}}, {{0x8a885d04, 0x1ceb, 0x11c9, {0x9f, 0xe8, 0x08, 0x00, 0x2b, 0x10, 0x48, 0x60}}, {2, 0}}, (RPC_DISPATCH_TABLE *) &drsuapi_v4_0_DispatchTable, 0, 0, 0, &drsuapi_ServerInfo, 0x04000000};
 static const RPC_CLIENT_INTERFACE drsuapi___RpcClientInterface = {sizeof(RPC_CLIENT_INTERFACE), {{0xe3514235, 0x4b06, 0x11d1, {0xab, 0x04, 0x00, 0xc0, 0x4f, 0xc2, 0xdc, 0xd2}}, {4, 0}}, {{0x8a885d04, 0x1ceb, 0x11c9, {0x9f, 0xe8, 0x08, 0x00, 0x2b, 0x10, 0x48, 0x60}}, {2, 0}}, 0, 0, 0, 0, 0, 0x00000000};
+RPC_IF_HANDLE
+	drsuapi_v4_0_s_ifspec = (RPC_IF_HANDLE) &drsuapi___RpcServerInterface,
+	drsuapi_v4_0_c_ifspec = (RPC_IF_HANDLE) &drsuapi___RpcClientInterface;
+static const NDR_RUNDOWN RundownRoutines[] = {SRV_DRS_HANDLE_rundown};
+static const SERVER_ROUTINE drsuapi_ServerRoutineTable[] = {
+	(SERVER_ROUTINE) SRV_IDL_DRSBind,
+	(SERVER_ROUTINE) SRV_IDL_DRSUnbind,
+	(SERVER_ROUTINE) SRV_OpnumNotImplemented,
+	(SERVER_ROUTINE) SRV_IDL_DRSGetNCChanges,
+	(SERVER_ROUTINE) SRV_IDL_DRSUpdateRefs,
+	(SERVER_ROUTINE) SRV_OpnumNotImplemented,
+	(SERVER_ROUTINE) SRV_OpnumNotImplemented,
+	(SERVER_ROUTINE) SRV_OpnumNotImplemented,
+	(SERVER_ROUTINE) SRV_OpnumNotImplemented,
+	(SERVER_ROUTINE) SRV_OpnumNotImplemented,
+	(SERVER_ROUTINE) SRV_OpnumNotImplemented,
+	(SERVER_ROUTINE) SRV_OpnumNotImplemented,
+	(SERVER_ROUTINE) SRV_IDL_DRSCrackNamesNotImplemented,
+	(SERVER_ROUTINE) SRV_OpnumNotImplemented,
+	(SERVER_ROUTINE) SRV_OpnumNotImplemented,
+	(SERVER_ROUTINE) SRV_OpnumNotImplemented,
+	(SERVER_ROUTINE) SRV_IDL_DRSDomainControllerInfoNotImplemented
+};
 static RPC_BINDING_HANDLE drsuapi__MIDL_AutoBindHandle;
-static const MIDL_STUB_DESC drsuapi_StubDesc = {(void *) &drsuapi___RpcClientInterface, MIDL_user_allocate, MIDL_user_free, &drsuapi__MIDL_AutoBindHandle, 0, 0, 0, 0, ms2Ddrsr__MIDL_TypeFormatString.Format, 1, 0x60000, 0, 0x8000253, 0, 0, 0, 0x1, 0, 0, 0};
+static const MIDL_STUB_DESC
+	drsuapi_s_StubDesc = {(void *) &drsuapi___RpcServerInterface, MIDL_user_allocate, MIDL_user_free, 0, RundownRoutines, 0, 0, 0, ms2Ddrsr__MIDL_TypeFormatString.Format, 1, 0x60000, 0, 0x8000253, 0, 0, 0, 0x1, 0, 0, 0},
+	drsuapi_c_StubDesc = {(void *) &drsuapi___RpcClientInterface, MIDL_user_allocate, MIDL_user_free, &drsuapi__MIDL_AutoBindHandle, 0, 0, 0, 0, ms2Ddrsr__MIDL_TypeFormatString.Format, 1, 0x60000, 0, 0x8000253, 0, 0, 0, 0x1, 0, 0, 0};
+static const MIDL_SERVER_INFO drsuapi_ServerInfo = {&drsuapi_s_StubDesc, drsuapi_ServerRoutineTable, ms2Ddrsr__MIDL_ProcFormatString.Format, drsuapi_FormatStringOffsetTable, 0, 0, 0, 0};
 static const MIDL_TYPE_PICKLING_INFO __MIDL_TypePicklingInfo = {0x33205054, 0x3, 0, 0, 0,};
 
+#ifdef _M_X64
 ULONG IDL_DRSBind(handle_t rpc_handle, UUID *puuidClientDsa, DRS_EXTENSIONS *pextClient, DRS_EXTENSIONS **ppextServer, DRS_HANDLE *phDrs)
 {
-    return (ULONG) NdrClientCall2((PMIDL_STUB_DESC) &drsuapi_StubDesc, (PFORMAT_STRING) &ms2Ddrsr__MIDL_ProcFormatString.Format[0], rpc_handle, puuidClientDsa, pextClient, ppextServer, phDrs).Simple;
+    return (ULONG) NdrClientCall2((PMIDL_STUB_DESC) &drsuapi_c_StubDesc, (PFORMAT_STRING) &ms2Ddrsr__MIDL_ProcFormatString.Format[0], rpc_handle, puuidClientDsa, pextClient, ppextServer, phDrs).Simple;
 }
 
 ULONG IDL_DRSUnbind(DRS_HANDLE *phDrs)
 {
-	return (ULONG) NdrClientCall2((PMIDL_STUB_DESC) &drsuapi_StubDesc, (PFORMAT_STRING) &ms2Ddrsr__MIDL_ProcFormatString.Format[60], phDrs).Simple;
+	return (ULONG) NdrClientCall2((PMIDL_STUB_DESC) &drsuapi_c_StubDesc, (PFORMAT_STRING) &ms2Ddrsr__MIDL_ProcFormatString.Format[60], phDrs).Simple;
 }
 
 ULONG IDL_DRSGetNCChanges(DRS_HANDLE hDrs, DWORD dwInVersion, DRS_MSG_GETCHGREQ *pmsgIn, DWORD *pdwOutVersion, DRS_MSG_GETCHGREPLY *pmsgOut)
 {
-    return (ULONG) NdrClientCall2((PMIDL_STUB_DESC)&drsuapi_StubDesc, (PFORMAT_STRING) &ms2Ddrsr__MIDL_ProcFormatString.Format[104], hDrs, dwInVersion, pmsgIn, pdwOutVersion, pmsgOut).Simple;
+    return (ULONG) NdrClientCall2((PMIDL_STUB_DESC) &drsuapi_c_StubDesc, (PFORMAT_STRING) &ms2Ddrsr__MIDL_ProcFormatString.Format[134], hDrs, dwInVersion, pmsgIn, pdwOutVersion, pmsgOut).Simple;
 }
 
 ULONG IDL_DRSCrackNames(DRS_HANDLE hDrs, DWORD dwInVersion, DRS_MSG_CRACKREQ *pmsgIn, DWORD *pdwOutVersion, DRS_MSG_CRACKREPLY *pmsgOut)
 {
-	return (ULONG) NdrClientCall2((PMIDL_STUB_DESC) &drsuapi_StubDesc, (PFORMAT_STRING) &ms2Ddrsr__MIDL_ProcFormatString.Format[172], hDrs, dwInVersion, pmsgIn, pdwOutVersion, pmsgOut).Simple;
+	return (ULONG) NdrClientCall2((PMIDL_STUB_DESC) &drsuapi_c_StubDesc, (PFORMAT_STRING) &ms2Ddrsr__MIDL_ProcFormatString.Format[468], hDrs, dwInVersion, pmsgIn, pdwOutVersion, pmsgOut).Simple;
 }
 
 ULONG IDL_DRSDomainControllerInfo(DRS_HANDLE hDrs, DWORD dwInVersion, DRS_MSG_DCINFOREQ *pmsgIn, DWORD *pdwOutVersion, DRS_MSG_DCINFOREPLY *pmsgOut)
 {
-	return (ULONG) NdrClientCall2((PMIDL_STUB_DESC) &drsuapi_StubDesc, (PFORMAT_STRING) &ms2Ddrsr__MIDL_ProcFormatString.Format[240], hDrs, dwInVersion, pmsgIn, pdwOutVersion, pmsgOut).Simple;
+	return (ULONG) NdrClientCall2((PMIDL_STUB_DESC) &drsuapi_c_StubDesc, (PFORMAT_STRING) &ms2Ddrsr__MIDL_ProcFormatString.Format[626], hDrs, dwInVersion, pmsgIn, pdwOutVersion, pmsgOut).Simple;
 }
 
 void DRS_MSG_GETCHGREPLY_V6_Free(handle_t _MidlEsHandle, DRS_MSG_GETCHGREPLY_V6 * _pType)
 {
-    NdrMesTypeFree2(_MidlEsHandle, (PMIDL_TYPE_PICKLING_INFO) &__MIDL_TypePicklingInfo, &drsuapi_StubDesc, (PFORMAT_STRING) &ms2Ddrsr__MIDL_TypeFormatString.Format[560], _pType);
+    NdrMesTypeFree2(_MidlEsHandle, (PMIDL_TYPE_PICKLING_INFO) &__MIDL_TypePicklingInfo, &drsuapi_c_StubDesc, (PFORMAT_STRING) &ms2Ddrsr__MIDL_TypeFormatString.Format[560], _pType);
 }
 
 void DRS_MSG_CRACKREPLY_V1_Free(handle_t _MidlEsHandle, DRS_MSG_CRACKREPLY_V1 * _pType)
 {
-    NdrMesTypeFree2(_MidlEsHandle, (PMIDL_TYPE_PICKLING_INFO ) &__MIDL_TypePicklingInfo, &drsuapi_StubDesc, (PFORMAT_STRING) &ms2Ddrsr__MIDL_TypeFormatString.Format[682], _pType);
+    NdrMesTypeFree2(_MidlEsHandle, (PMIDL_TYPE_PICKLING_INFO ) &__MIDL_TypePicklingInfo, &drsuapi_c_StubDesc, (PFORMAT_STRING) &ms2Ddrsr__MIDL_TypeFormatString.Format[682], _pType);
 }
 
 void DRS_MSG_DCINFOREPLY_V2_Free(handle_t _MidlEsHandle, DRS_MSG_DCINFOREPLY_V2 * _pType)
 {
-    NdrMesTypeFree2(_MidlEsHandle, (PMIDL_TYPE_PICKLING_INFO  )&__MIDL_TypePicklingInfo, &drsuapi_StubDesc, (PFORMAT_STRING) &ms2Ddrsr__MIDL_TypeFormatString.Format[792], _pType);
+    NdrMesTypeFree2(_MidlEsHandle, (PMIDL_TYPE_PICKLING_INFO  )&__MIDL_TypePicklingInfo, &drsuapi_c_StubDesc, (PFORMAT_STRING) &ms2Ddrsr__MIDL_TypeFormatString.Format[792], _pType);
 }
 
 static const ms2Ddrsr_MIDL_PROC_FORMAT_STRING ms2Ddrsr__MIDL_ProcFormatString = {0, {
 	0x00, 0x48, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x30, 0x00, 0x32, 0x00, 0x00, 0x00, 0x44, 0x00, 0x40, 0x00, 0x47, 0x05, 0x0a, 0x07, 0x01, 0x00, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x0a, 0x00,
 	0x08, 0x00, 0x2c, 0x03, 0x0b, 0x00, 0x10, 0x00, 0x30, 0x03, 0x13, 0x20, 0x18, 0x00, 0x58, 0x03, 0x10, 0x01, 0x20, 0x00, 0x60, 0x03, 0x70, 0x00, 0x28, 0x00, 0x08, 0x00, 0x00, 0x48, 0x00, 0x00,
 	0x00, 0x00, 0x01, 0x00, 0x10, 0x00, 0x30, 0xe0, 0x00, 0x00, 0x00, 0x00, 0x38, 0x00, 0x40, 0x00, 0x44, 0x02, 0x0a, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x18, 0x01, 0x00, 0x00,
-	0x68, 0x03, 0x70, 0x00, 0x08, 0x00, 0x08, 0x00, 0x00, 0x48, 0x00, 0x00, 0x00, 0x00, 0x03, 0x00, 0x30, 0x00, 0x30, 0x40, 0x00, 0x00, 0x00, 0x00, 0x2c, 0x00, 0x24, 0x00, 0x47, 0x06, 0x0a, 0x07,
-	0x01, 0x00, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x08, 0x00, 0x00, 0x00, 0x6c, 0x03, 0x48, 0x00, 0x08, 0x00, 0x08, 0x00, 0x0b, 0x01, 0x10, 0x00, 0x74, 0x03, 0x50, 0x21, 0x18, 0x00, 0x08, 0x00,
-	0x13, 0x01, 0x20, 0x00, 0x1e, 0x05, 0x70, 0x00, 0x28, 0x00, 0x08, 0x00, 0x00, 0x48, 0x00, 0x00, 0x00, 0x00, 0x0c, 0x00, 0x30, 0x00, 0x30, 0x40, 0x00, 0x00, 0x00, 0x00, 0x2c, 0x00, 0x24, 0x00,
-	0x47, 0x06, 0x0a, 0x07, 0x01, 0x00, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x08, 0x00, 0x00, 0x00, 0x6c, 0x03, 0x48, 0x00, 0x08, 0x00, 0x08, 0x00, 0x0b, 0x01, 0x10, 0x00, 0x44, 0x06, 0x50, 0x21,
-	0x18, 0x00, 0x08, 0x00, 0x13, 0x21, 0x20, 0x00, 0x96, 0x06, 0x70, 0x00, 0x28, 0x00, 0x08, 0x00, 0x00, 0x48, 0x00, 0x00, 0x00, 0x00, 0x10, 0x00, 0x30, 0x00, 0x30, 0x40, 0x00, 0x00, 0x00, 0x00,
-	0x2c, 0x00, 0x24, 0x00, 0x47, 0x06, 0x0a, 0x07, 0x01, 0x00, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x08, 0x00, 0x00, 0x00, 0x6c, 0x03, 0x48, 0x00, 0x08, 0x00, 0x08, 0x00, 0x0b, 0x01, 0x10, 0x00,
-	0xb0, 0x06, 0x50, 0x21, 0x18, 0x00, 0x08, 0x00, 0x13, 0x41, 0x20, 0x00, 0xda, 0x06, 0x70, 0x00, 0x28, 0x00, 0x08, 0x00, 0x00,
+	0x68, 0x03, 0x70, 0x00, 0x08, 0x00, 0x08, 0x00, 0x00, 0x48, 0x00, 0x00, 0x00, 0x00, 0x02, 0x00, 0x08, 0x00, 0x32, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x40, 0x00, 0x0a, 0x01, 0x00, 0x00,
+	0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x48, 0x00, 0x00, 0x00, 0x00, 0x03, 0x00, 0x30, 0x00, 0x30, 0x40, 0x00, 0x00, 0x00, 0x00, 0x2c, 0x00, 0x24, 0x00, 0x47, 0x06, 0x0a, 0x07, 0x01, 0x00,
+	0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x08, 0x00, 0x00, 0x00, 0x6c, 0x03, 0x48, 0x00, 0x08, 0x00, 0x08, 0x00, 0x0b, 0x01, 0x10, 0x00, 0x74, 0x03, 0x50, 0x21, 0x18, 0x00, 0x08, 0x00, 0x13, 0x01,
+	0x20, 0x00, 0x28, 0x04, 0x70, 0x00, 0x28, 0x00, 0x08, 0x00, 0x00, 0x48, 0x00, 0x00, 0x00, 0x00, 0x04, 0x00, 0x20, 0x00, 0x30, 0x40, 0x00, 0x00, 0x00, 0x00, 0x2c, 0x00, 0x08, 0x00, 0x46, 0x04,
+	0x0a, 0x05, 0x00, 0x00, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x08, 0x00, 0x00, 0x00, 0x6c, 0x03, 0x48, 0x00, 0x08, 0x00, 0x08, 0x00, 0x0b, 0x01, 0x10, 0x00, 0x42, 0x04, 0x70, 0x00, 0x18, 0x00,
+	0x08, 0x00, 0x00, 0x48, 0x00, 0x00, 0x00, 0x00, 0x05, 0x00, 0x08, 0x00, 0x32, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x40, 0x00, 0x0a, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+	0x00, 0x48, 0x00, 0x00, 0x00, 0x00, 0x06, 0x00, 0x08, 0x00, 0x32, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x40, 0x00, 0x0a, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x48,
+	0x00, 0x00, 0x00, 0x00, 0x07, 0x00, 0x08, 0x00, 0x32, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x40, 0x00, 0x0a, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x48, 0x00, 0x00,
+	0x00, 0x00, 0x08, 0x00, 0x08, 0x00, 0x32, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x40, 0x00, 0x0a, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x48, 0x00, 0x00, 0x00, 0x00,
+	0x09, 0x00, 0x08, 0x00, 0x32, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x40, 0x00, 0x0a, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x48, 0x00, 0x00, 0x00, 0x00, 0x0a, 0x00,
+	0x08, 0x00, 0x32, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x40, 0x00, 0x0a, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x48, 0x00, 0x00, 0x00, 0x00, 0x0b, 0x00, 0x08, 0x00,
+	0x32, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x40, 0x00, 0x0a, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x48, 0x00, 0x00, 0x00, 0x00, 0x0c, 0x00, 0x30, 0x00, 0x30, 0x40,
+	0x00, 0x00, 0x00, 0x00, 0x2c, 0x00, 0x24, 0x00, 0x47, 0x06, 0x0a, 0x07, 0x01, 0x00, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x08, 0x00, 0x00, 0x00, 0x6c, 0x03, 0x48, 0x00, 0x08, 0x00, 0x08, 0x00,
+	0x0b, 0x01, 0x10, 0x00, 0x76, 0x04, 0x50, 0x21, 0x18, 0x00, 0x08, 0x00, 0x13, 0x21, 0x20, 0x00, 0xc8, 0x04, 0x70, 0x00, 0x28, 0x00, 0x08, 0x00, 0x00, 0x48, 0x00, 0x00, 0x00, 0x00, 0x0d, 0x00,
+	0x08, 0x00, 0x32, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x40, 0x00, 0x0a, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x48, 0x00, 0x00, 0x00, 0x00, 0x0e, 0x00, 0x08, 0x00,
+	0x32, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x40, 0x00, 0x0a, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x48, 0x00, 0x00, 0x00, 0x00, 0x0f, 0x00, 0x08, 0x00, 0x32, 0x00,
+	0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x40, 0x00, 0x0a, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x48, 0x00, 0x00, 0x00, 0x00, 0x10, 0x00, 0x30, 0x00, 0x30, 0x40, 0x00, 0x00,
+	0x00, 0x00, 0x2c, 0x00, 0x24, 0x00, 0x47, 0x06, 0x0a, 0x07, 0x01, 0x00, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x08, 0x00, 0x00, 0x00, 0x6c, 0x03, 0x48, 0x00, 0x08, 0x00, 0x08, 0x00, 0x0b, 0x01,
+	0x10, 0x00, 0xe2, 0x04, 0x50, 0x21, 0x18, 0x00, 0x08, 0x00, 0x13, 0x41, 0x20, 0x00, 0x0c, 0x05, 0x70, 0x00, 0x28, 0x00, 0x08, 0x00, 0x00
 }};
 static const ms2Ddrsr_MIDL_TYPE_FORMAT_STRING ms2Ddrsr__MIDL_TypeFormatString = {0, {
 	0x00, 0x00, 0x1d, 0x00, 0x08, 0x00, 0x01, 0x5b, 0x15, 0x03, 0x10, 0x00, 0x08, 0x06, 0x06, 0x4c, 0x00, 0xf1, 0xff, 0x5b, 0x15, 0x07, 0x18, 0x00, 0x0b, 0x0b, 0x0b, 0x5b, 0xb7, 0x08, 0x00, 0x00,
-	0x00, 0x00, 0x00, 0x00, 0x10, 0x00, 0xb7, 0x08, 0x00, 0x00, 0x00, 0x00, 0x10, 0x27, 0x00, 0x00, 0x1b, 0x00, 0x01, 0x00, 0x19, 0x00, 0x00, 0x00, 0x01, 0x00, 0x02, 0x5b, 0x1a, 0x03, 0x10, 0x00,
-	0x00, 0x00, 0x0a, 0x00, 0x4c, 0x00, 0xe0, 0xff, 0x40, 0x36, 0x5c, 0x5b, 0x12, 0x00, 0xe2, 0xff, 0x1a, 0x03, 0x18, 0x00, 0x00, 0x00, 0x00, 0x00, 0x08, 0x40, 0x4c, 0x00, 0xe0, 0xff, 0x5c, 0x5b,
-	0x21, 0x03, 0x00, 0x00, 0x19, 0x00, 0x00, 0x00, 0x01, 0x00, 0xff, 0xff, 0xff, 0xff, 0x00, 0x00, 0x4c, 0x00, 0xde, 0xff, 0x5c, 0x5b, 0x1a, 0x03, 0x10, 0x00, 0x00, 0x00, 0x0a, 0x00, 0x4c, 0x00,
-	0x9c, 0xff, 0x40, 0x36, 0x5c, 0x5b, 0x12, 0x00, 0xd8, 0xff, 0xb7, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x10, 0x00, 0x1d, 0x00, 0x1c, 0x00, 0x02, 0x5b, 0x15, 0x00, 0x1c, 0x00, 0x4c, 0x00,
-	0xf4, 0xff, 0x5c, 0x5b, 0x1b, 0x01, 0x02, 0x00, 0x09, 0x57, 0xfc, 0xff, 0x01, 0x00, 0x05, 0x5b, 0x17, 0x03, 0x38, 0x00, 0xf0, 0xff, 0x08, 0x08, 0x4c, 0x00, 0x4e, 0xff, 0x4c, 0x00, 0xdc, 0xff,
-	0x08, 0x5b, 0xb7, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x10, 0x00, 0x15, 0x07, 0x20, 0x00, 0x4c, 0x00, 0x36, 0xff, 0x0b, 0x0b, 0x5c, 0x5b, 0x1b, 0x07, 0x20, 0x00, 0x09, 0x00, 0xf8, 0xff,
-	0x01, 0x00, 0x4c, 0x00, 0xe8, 0xff, 0x5c, 0x5b, 0x1a, 0x07, 0x10, 0x00, 0xec, 0xff, 0x00, 0x00, 0x08, 0x08, 0x4c, 0x00, 0xce, 0xff, 0x08, 0x5b, 0xb7, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-	0x10, 0x00, 0xb7, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0xa0, 0x00, 0xb7, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x90, 0x01, 0x1a, 0x03, 0x10, 0x00, 0x00, 0x00, 0x0a, 0x00, 0x4c, 0x00,
-	0xec, 0xff, 0x40, 0x36, 0x5c, 0x5b, 0x12, 0x00, 0x08, 0xff, 0x21, 0x03, 0x00, 0x00, 0x19, 0x00, 0x00, 0x00, 0x01, 0x00, 0xff, 0xff, 0xff, 0xff, 0x00, 0x00, 0x4c, 0x00, 0xda, 0xff, 0x5c, 0x5b,
-	0x1a, 0x03, 0x10, 0x00, 0x00, 0x00, 0x0a, 0x00, 0x4c, 0x00, 0xb8, 0xff, 0x40, 0x36, 0x5c, 0x5b, 0x12, 0x00, 0xd8, 0xff, 0x1a, 0x03, 0x18, 0x00, 0x00, 0x00, 0x00, 0x00, 0x08, 0x40, 0x4c, 0x00,
-	0xe0, 0xff, 0x5c, 0x5b, 0x21, 0x03, 0x00, 0x00, 0x19, 0x00, 0x00, 0x00, 0x01, 0x00, 0xff, 0xff, 0xff, 0xff, 0x00, 0x00, 0x4c, 0x00, 0xde, 0xff, 0x5c, 0x5b, 0x1a, 0x03, 0x10, 0x00, 0x00, 0x00,
-	0x0a, 0x00, 0x4c, 0x00, 0x74, 0xff, 0x40, 0x36, 0x5c, 0x5b, 0x12, 0x00, 0xd8, 0xff, 0x1a, 0x03, 0x20, 0x00, 0x00, 0x00, 0x0a, 0x00, 0x36, 0x08, 0x40, 0x4c, 0x00, 0xdf, 0xff, 0x5b, 0x12, 0x00,
-	0x10, 0xff, 0xb7, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x10, 0x00, 0x15, 0x07, 0x28, 0x00, 0x08, 0x40, 0x0b, 0x4c, 0x00, 0x53, 0xfe, 0x0b, 0x5c, 0x5b, 0x1b, 0x07, 0x28, 0x00, 0x09, 0x00,
-	0xf8, 0xff, 0x01, 0x00, 0x4c, 0x00, 0xe6, 0xff, 0x5c, 0x5b, 0x1a, 0x07, 0x08, 0x00, 0xec, 0xff, 0x00, 0x00, 0x4c, 0x00, 0xce, 0xff, 0x40, 0x5b, 0x1a, 0x03, 0x40, 0x00, 0x00, 0x00, 0x0c, 0x00,
-	0x36, 0x4c, 0x00, 0xab, 0xff, 0x08, 0x40, 0x36, 0x36, 0x5b, 0x12, 0x00, 0xec, 0xff, 0x12, 0x00, 0x18, 0xfe, 0x12, 0x00, 0xd6, 0xff, 0x15, 0x07, 0x30, 0x00, 0x0b, 0x4c, 0x00, 0xaf, 0xff, 0x5b,
-	0x1a, 0x07, 0x58, 0x00, 0x00, 0x00, 0x10, 0x00, 0x36, 0x08, 0x40, 0x4c, 0x00, 0x09, 0xff, 0x08, 0x40, 0x4c, 0x00, 0xe3, 0xff, 0x5b, 0x12, 0x00, 0x98, 0xfe, 0x21, 0x07, 0x00, 0x00, 0x19, 0x00,
-	0x94, 0x00, 0x01, 0x00, 0xff, 0xff, 0xff, 0xff, 0x00, 0x00, 0x4c, 0x00, 0xd4, 0xff, 0x5c, 0x5b, 0x1a, 0x07, 0xa8, 0x00, 0x00, 0x00, 0x28, 0x00, 0x4c, 0x00, 0xce, 0xfd, 0x4c, 0x00, 0xca, 0xfd,
-	0x36, 0x4c, 0x00, 0xd1, 0xfd, 0x4c, 0x00, 0xcd, 0xfd, 0x36, 0x4c, 0x00, 0x2a, 0xfe, 0x08, 0x08, 0x08, 0x40, 0x36, 0x08, 0x08, 0x08, 0x4c, 0x00, 0x32, 0xfe, 0x36, 0x08, 0x40, 0x5b, 0x12, 0x00,
-	0x50, 0xfe, 0x12, 0x00, 0x84, 0xfe, 0x12, 0x00, 0x70, 0xff, 0x12, 0x00, 0xae, 0xff, 0x1a, 0x03, 0x18, 0x00, 0x00, 0x00, 0x08, 0x00, 0x08, 0x40, 0x36, 0x36, 0x5c, 0x5b, 0x12, 0x08, 0x25, 0x5c,
-	0x12, 0x08, 0x25, 0x5c, 0x21, 0x03, 0x00, 0x00, 0x19, 0x00, 0x00, 0x00, 0x01, 0x00, 0xff, 0xff, 0xff, 0xff, 0x00, 0x00, 0x4c, 0x00, 0xd8, 0xff, 0x5c, 0x5b, 0x1a, 0x03, 0x10, 0x00, 0x00, 0x00,
-	0x06, 0x00, 0x08, 0x40, 0x36, 0x5b, 0x12, 0x00, 0xdc, 0xff, 0x1a, 0x03, 0x08, 0x00, 0x00, 0x00, 0x04, 0x00, 0x36, 0x5b, 0x12, 0x00, 0xe4, 0xff, 0xb7, 0x08, 0x00, 0x00, 0x00, 0x00, 0x10, 0x27,
-	0x00, 0x00, 0x1a, 0x03, 0x88, 0x00, 0x00, 0x00, 0x1e, 0x00, 0x36, 0x36, 0x36, 0x36, 0x36, 0x36, 0x36, 0x08, 0x08, 0x08, 0x4c, 0x00, 0x32, 0xfd, 0x4c, 0x00, 0x2e, 0xfd, 0x4c, 0x00, 0x2a, 0xfd,
-	0x4c, 0x00, 0x26, 0xfd, 0x40, 0x5b, 0x12, 0x08, 0x25, 0x5c, 0x12, 0x08, 0x25, 0x5c, 0x12, 0x08, 0x25, 0x5c, 0x12, 0x08, 0x25, 0x5c, 0x12, 0x08, 0x25, 0x5c, 0x12, 0x08, 0x25, 0x5c, 0x12, 0x08,
-	0x25, 0x5c, 0x21, 0x03, 0x00, 0x00, 0x19, 0x00, 0x00, 0x00, 0x01, 0x00, 0xff, 0xff, 0xff, 0xff, 0x00, 0x00, 0x4c, 0x00, 0xae, 0xff, 0x5c, 0x5b, 0x1a, 0x03, 0x10, 0x00, 0x00, 0x00, 0x0a, 0x00,
-	0x4c, 0x00, 0x96, 0xff, 0x40, 0x36, 0x5c, 0x5b, 0x12, 0x00, 0xd8, 0xff, 0x12, 0x00, 0xda, 0xfc, 0x12, 0x00, 0x18, 0x00, 0xb7, 0x08, 0x01, 0x00, 0x00, 0x00, 0x10, 0x27, 0x00, 0x00, 0x1b, 0x00,
-	0x01, 0x00, 0x09, 0x00, 0xfc, 0xff, 0x01, 0x00, 0x02, 0x5b, 0x1a, 0x03, 0x04, 0x00, 0xf0, 0xff, 0x00, 0x00, 0x4c, 0x00, 0xe0, 0xff, 0x5c, 0x5b, 0x11, 0x14, 0xd6, 0xff, 0x11, 0x04, 0x02, 0x00,
-	0x30, 0xa0, 0x00, 0x00, 0x11, 0x04, 0x02, 0x00, 0x30, 0xe1, 0x00, 0x00, 0x30, 0x41, 0x00, 0x00, 0x11, 0x00, 0x02, 0x00, 0x2b, 0x09, 0x29, 0x00, 0x08, 0x00, 0x01, 0x00, 0x02, 0x00, 0xa8, 0x00,
-	0x05, 0x00, 0x04, 0x00, 0x00, 0x00, 0xba, 0x00, 0x05, 0x00, 0x00, 0x00, 0xd0, 0x00, 0x07, 0x00, 0x00, 0x00, 0xf2, 0x00, 0x08, 0x00, 0x00, 0x00, 0x10, 0x01, 0x0a, 0x00, 0x00, 0x00, 0x40, 0x01,
-	0xff, 0xff, 0xb7, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x10, 0x00, 0x15, 0x07, 0x18, 0x00, 0x4c, 0x00, 0x56, 0xfc, 0x0b, 0x5b, 0x1b, 0x07, 0x18, 0x00, 0x09, 0x00, 0xf8, 0xff, 0x01, 0x00,
-	0x4c, 0x00, 0xea, 0xff, 0x5c, 0x5b, 0x1a, 0x07, 0x10, 0x00, 0xec, 0xff, 0x00, 0x00, 0x08, 0x08, 0x4c, 0x00, 0xd0, 0xff, 0x08, 0x5b, 0xb7, 0x08, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x10, 0x00,
-	0x1b, 0x03, 0x04, 0x00, 0x09, 0x00, 0xfc, 0xff, 0x01, 0x00, 0x08, 0x5b, 0x1a, 0x03, 0x0c, 0x00, 0xf0, 0xff, 0x00, 0x00, 0x08, 0x08, 0x4c, 0x00, 0xde, 0xff, 0x5c, 0x5b, 0x1a, 0x07, 0x70, 0x00,
-	0x00, 0x00, 0x1a, 0x00, 0x4c, 0x00, 0x02, 0xfc, 0x4c, 0x00, 0xfe, 0xfb, 0x36, 0x4c, 0x00, 0x05, 0xfc, 0x36, 0x36, 0x4c, 0x00, 0x61, 0xfc, 0x08, 0x08, 0x08, 0x08, 0x5b, 0x11, 0x00, 0x92, 0xfc,
-	0x12, 0x00, 0xa4, 0xff, 0x12, 0x00, 0xc6, 0xff, 0xb7, 0x08, 0x01, 0x00, 0x00, 0x00, 0x00, 0x01, 0x00, 0x00, 0x1a, 0x03, 0x04, 0x00, 0x08, 0xff, 0x00, 0x00, 0x4c, 0x00, 0xec, 0xff, 0x5c, 0x5b,
-	0x1a, 0x07, 0x88, 0x00, 0x00, 0x00, 0x0c, 0x00, 0x4c, 0x00, 0xbe, 0xfb, 0x36, 0x4c, 0x00, 0xad, 0xff, 0x5b, 0x11, 0x00, 0xde, 0xff, 0x15, 0x07, 0x08, 0x00, 0x0b, 0x5b, 0x1a, 0x07, 0x60, 0x00,
-	0x00, 0x00, 0x1a, 0x00, 0x4c, 0x00, 0xa2, 0xfb, 0x4c, 0x00, 0x9e, 0xfb, 0x36, 0x4c, 0x00, 0xa5, 0xfb, 0x36, 0x08, 0x08, 0x08, 0x08, 0x4c, 0x00, 0xde, 0xff, 0x5c, 0x5b, 0x11, 0x00, 0x32, 0xfc,
-	0x12, 0x00, 0x44, 0xff, 0x1a, 0x07, 0xa8, 0x00, 0x00, 0x00, 0x12, 0x00, 0x4c, 0x00, 0x7a, 0xfb, 0x36, 0x4c, 0x00, 0x69, 0xff, 0x36, 0x36, 0x4c, 0x00, 0xdd, 0xfb, 0x5b, 0x11, 0x00, 0x94, 0xff,
-	0x12, 0x00, 0x4a, 0xff, 0x12, 0x00, 0x46, 0xff, 0x1a, 0x07, 0x80, 0x00, 0x00, 0x00, 0x20, 0x00, 0x4c, 0x00, 0x56, 0xfb, 0x4c, 0x00, 0x52, 0xfb, 0x36, 0x4c, 0x00, 0x59, 0xfb, 0x36, 0x08, 0x08,
-	0x08, 0x08, 0x4c, 0x00, 0x92, 0xff, 0x36, 0x36, 0x4c, 0x00, 0xac, 0xfb, 0x5c, 0x5b, 0x11, 0x00, 0xe0, 0xfb, 0x12, 0x00, 0xf2, 0xfe, 0x12, 0x00, 0x14, 0xff, 0x12, 0x00, 0x10, 0xff, 0x1a, 0x07,
-	0x88, 0x00, 0x00, 0x00, 0x22, 0x00, 0x4c, 0x00, 0x20, 0xfb, 0x4c, 0x00, 0x1c, 0xfb, 0x36, 0x4c, 0x00, 0x23, 0xfb, 0x36, 0x08, 0x08, 0x08, 0x08, 0x4c, 0x00, 0x5c, 0xff, 0x36, 0x36, 0x4c, 0x00,
-	0x76, 0xfb, 0x08, 0x40, 0x5c, 0x5b, 0x11, 0x00, 0xa8, 0xfb, 0x12, 0x00, 0xba, 0xfe, 0x12, 0x00, 0xdc, 0xfe, 0x12, 0x00, 0xd8, 0xfe, 0x11, 0x0c, 0x08, 0x5c, 0x11, 0x00, 0x02, 0x00, 0x2b, 0x09,
-	0x29, 0x54, 0x18, 0x00, 0x01, 0x00, 0x02, 0x00, 0xa8, 0x00, 0x05, 0x00, 0x01, 0x00, 0x00, 0x00, 0x1c, 0x00, 0x02, 0x00, 0x00, 0x00, 0x64, 0x00, 0x06, 0x00, 0x00, 0x00, 0xf4, 0xfc, 0x07, 0x00,
-	0x00, 0x00, 0x66, 0x00, 0x09, 0x00, 0x00, 0x00, 0xba, 0x00, 0xff, 0xff, 0x1a, 0x07, 0x90, 0x00, 0x00, 0x00, 0x20, 0x00, 0x4c, 0x00, 0xb2, 0xfa, 0x4c, 0x00, 0xae, 0xfa, 0x36, 0x4c, 0x00, 0xb5,
-	0xfa, 0x4c, 0x00, 0xb1, 0xfa, 0x36, 0x4c, 0x00, 0x0e, 0xfb, 0x08, 0x08, 0x08, 0x40, 0x36, 0x08, 0x40, 0x5b, 0x12, 0x00, 0x3c, 0xfb, 0x12, 0x00, 0x4e, 0xfe, 0x12, 0x00, 0x5c, 0xfc, 0x1b, 0x00,
-	0x01, 0x00, 0x19, 0x00, 0x04, 0x00, 0x01, 0x00, 0x02, 0x5b, 0x1a, 0x03, 0x10, 0x00, 0x00, 0x00, 0x06, 0x00, 0x08, 0x08, 0x36, 0x5b, 0x12, 0x00, 0xe6, 0xff, 0x1a, 0x03, 0x10, 0x00, 0x00, 0x00,
-	0x00, 0x00, 0x4c, 0x00, 0xe6, 0xff, 0x5c, 0x5b, 0x1a, 0x03, 0x18, 0x00, 0x00, 0x00, 0x00, 0x00, 0x08, 0x0d, 0x4c, 0x00, 0xd6, 0xff, 0x5c, 0x5b, 0xb7, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-	0x10, 0x00, 0x15, 0x07, 0x48, 0x00, 0x0b, 0x4c, 0x00, 0xe3, 0xfb, 0x08, 0x08, 0x08, 0x40, 0x0b, 0x5c, 0x5b, 0x1a, 0x07, 0x70, 0x00, 0x00, 0x00, 0x10, 0x00, 0x36, 0x08, 0x40, 0x4c, 0x00, 0x37,
-	0xfb, 0x08, 0x40, 0x4c, 0x00, 0xdd, 0xff, 0x5b, 0x12, 0x00, 0xc6, 0xfa, 0x21, 0x07, 0x00, 0x00, 0x19, 0x00, 0x94, 0x00, 0x01, 0x00, 0xff, 0xff, 0xff, 0xff, 0x00, 0x00, 0x4c, 0x00, 0xd4, 0xff,
-	0x5c, 0x5b, 0x1a, 0x07, 0xa8, 0x00, 0x00, 0x00, 0x28, 0x00, 0x4c, 0x00, 0xfc, 0xf9, 0x4c, 0x00, 0xf8, 0xf9, 0x36, 0x4c, 0x00, 0xff, 0xf9, 0x4c, 0x00, 0xfb, 0xf9, 0x36, 0x4c, 0x00, 0x58, 0xfa,
-	0x08, 0x08, 0x08, 0x40, 0x36, 0x08, 0x08, 0x08, 0x4c, 0x00, 0x8e, 0xff, 0x36, 0x08, 0x40, 0x5b, 0x12, 0x00, 0x7e, 0xfa, 0x12, 0x00, 0xb2, 0xfa, 0x12, 0x00, 0x9e, 0xfb, 0x12, 0x00, 0xae, 0xff,
-	0x11, 0x00, 0x02, 0x00, 0x2b, 0x09, 0x29, 0x00, 0x08, 0x00, 0x01, 0x00, 0x02, 0x00, 0x20, 0x00, 0x01, 0x00, 0x01, 0x00, 0x00, 0x00, 0x24, 0x00, 0xff, 0xff, 0xb7, 0x08, 0x01, 0x00, 0x00, 0x00,
-	0x10, 0x27, 0x00, 0x00, 0x21, 0x03, 0x00, 0x00, 0x19, 0x00, 0x14, 0x00, 0x01, 0x00, 0xff, 0xff, 0xff, 0xff, 0x00, 0x00, 0x12, 0x08, 0x25, 0x5c, 0x5c, 0x5b, 0x1a, 0x03, 0x20, 0x00, 0x00, 0x00,
-	0x0e, 0x00, 0x08, 0x08, 0x08, 0x08, 0x08, 0x4c, 0x00, 0xd1, 0xff, 0x36, 0x5c, 0x5b, 0x12, 0x00, 0xd4, 0xff, 0x11, 0x04, 0x02, 0x00, 0x2b, 0x09, 0x29, 0x54, 0x18, 0x00, 0x01, 0x00, 0x02, 0x00,
-	0x08, 0x00, 0x01, 0x00, 0x01, 0x00, 0x00, 0x00, 0x02, 0xfc, 0xff, 0xff, 0x11, 0x00, 0x02, 0x00, 0x2b, 0x09, 0x29, 0x00, 0x08, 0x00, 0x01, 0x00, 0x02, 0x00, 0x10, 0x00, 0x01, 0x00, 0x01, 0x00,
-	0x00, 0x00, 0x04, 0x00, 0xff, 0xff, 0x1a, 0x03, 0x10, 0x00, 0x00, 0x00, 0x06, 0x00, 0x36, 0x08, 0x40, 0x5b, 0x12, 0x08, 0x25, 0x5c, 0x11, 0x04, 0x02, 0x00, 0x2b, 0x09, 0x29, 0x54, 0x18, 0x00,
-	0x01, 0x00, 0x02, 0x00, 0x10, 0x00, 0x04, 0x00, 0x01, 0x00, 0x00, 0x00, 0x5a, 0x00, 0x02, 0x00, 0x00, 0x00, 0x26, 0xfc, 0x03, 0x00, 0x00, 0x00, 0xc2, 0x00, 0xff, 0xff, 0xff, 0xff, 0x04, 0x01,
-	0xff, 0xff, 0xb7, 0x08, 0x00, 0x00, 0x00, 0x00, 0x10, 0x27, 0x00, 0x00, 0x1a, 0x03, 0x30, 0x00, 0x00, 0x00, 0x0a, 0x00, 0x36, 0x36, 0x36, 0x36, 0x36, 0x08, 0x08, 0x5b, 0x12, 0x08, 0x25, 0x5c,
-	0x12, 0x08, 0x25, 0x5c, 0x12, 0x08, 0x25, 0x5c, 0x12, 0x08, 0x25, 0x5c, 0x12, 0x08, 0x25, 0x5c, 0x21, 0x03, 0x00, 0x00, 0x19, 0x00, 0x00, 0x00, 0x01, 0x00, 0xff, 0xff, 0xff, 0xff, 0x00, 0x00,
-	0x4c, 0x00, 0xca, 0xff, 0x5c, 0x5b, 0x1a, 0x03, 0x10, 0x00, 0x00, 0x00, 0x0a, 0x00, 0x4c, 0x00, 0xb2, 0xff, 0x40, 0x36, 0x5c, 0x5b, 0x12, 0x00, 0xd8, 0xff, 0xb7, 0x08, 0x00, 0x00, 0x00, 0x00,
-	0x10, 0x27, 0x00, 0x00, 0x1a, 0x03, 0x88, 0x00, 0x00, 0x00, 0x1e, 0x00, 0x36, 0x36, 0x36, 0x36, 0x36, 0x36, 0x36, 0x08, 0x08, 0x08, 0x08, 0x4c, 0x00, 0x8f, 0xf8, 0x4c, 0x00, 0x8b, 0xf8, 0x4c,
-	0x00, 0x87, 0xf8, 0x4c, 0x00, 0x83, 0xf8, 0x5b, 0x12, 0x08, 0x25, 0x5c, 0x12, 0x08, 0x25, 0x5c, 0x12, 0x08, 0x25, 0x5c, 0x12, 0x08, 0x25, 0x5c, 0x12, 0x08, 0x25, 0x5c, 0x12, 0x08, 0x25, 0x5c,
-	0x12, 0x08, 0x25, 0x5c, 0x21, 0x03, 0x00, 0x00, 0x19, 0x00, 0x00, 0x00, 0x01, 0x00, 0xff, 0xff, 0xff, 0xff, 0x00, 0x00, 0x4c, 0x00, 0xae, 0xff, 0x5c, 0x5b, 0x1a, 0x03, 0x10, 0x00, 0x00, 0x00,
-	0x0a, 0x00, 0x4c, 0x00, 0x96, 0xff, 0x40, 0x36, 0x5c, 0x5b, 0x12, 0x00, 0xd8, 0xff, 0xb7, 0x08, 0x00, 0x00, 0x00, 0x00, 0x10, 0x27, 0x00, 0x00, 0x1a, 0x03, 0x20, 0x00, 0x00, 0x00, 0x0a, 0x00,
-	0x08, 0x08, 0x08, 0x08, 0x08, 0x08, 0x36, 0x5b, 0x12, 0x08, 0x25, 0x5c, 0x21, 0x03, 0x00, 0x00, 0x19, 0x00, 0x00, 0x00, 0x01, 0x00, 0xff, 0xff, 0xff, 0xff, 0x00, 0x00, 0x4c, 0x00, 0xda, 0xff,
-	0x5c, 0x5b, 0x1a, 0x03, 0x10, 0x00, 0x00, 0x00, 0x0a, 0x00, 0x4c, 0x00, 0xc2, 0xff, 0x40, 0x36, 0x5c, 0x5b, 0x12, 0x00, 0xd8, 0xff, 0x00,
+    0x00, 0x00, 0x00, 0x00, 0x10, 0x00, 0xb7, 0x08, 0x00, 0x00, 0x00, 0x00, 0x10, 0x27, 0x00, 0x00, 0x1b, 0x00, 0x01, 0x00, 0x19, 0x00, 0x00, 0x00, 0x01, 0x00, 0x02, 0x5b, 0x1a, 0x03, 0x10, 0x00,
+    0x00, 0x00, 0x0a, 0x00, 0x4c, 0x00, 0xe0, 0xff, 0x40, 0x36, 0x5c, 0x5b, 0x12, 0x00, 0xe2, 0xff, 0x1a, 0x03, 0x18, 0x00, 0x00, 0x00, 0x00, 0x00, 0x08, 0x40, 0x4c, 0x00, 0xe0, 0xff, 0x5c, 0x5b,
+    0x21, 0x03, 0x00, 0x00, 0x19, 0x00, 0x00, 0x00, 0x01, 0x00, 0xff, 0xff, 0xff, 0xff, 0x00, 0x00, 0x4c, 0x00, 0xde, 0xff, 0x5c, 0x5b, 0x1a, 0x03, 0x10, 0x00, 0x00, 0x00, 0x0a, 0x00, 0x4c, 0x00,
+    0x9c, 0xff, 0x40, 0x36, 0x5c, 0x5b, 0x12, 0x00, 0xd8, 0xff, 0xb7, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x10, 0x00, 0x1d, 0x00, 0x1c, 0x00, 0x02, 0x5b, 0x15, 0x00, 0x1c, 0x00, 0x4c, 0x00,
+    0xf4, 0xff, 0x5c, 0x5b, 0x1b, 0x01, 0x02, 0x00, 0x09, 0x57, 0xfc, 0xff, 0x01, 0x00, 0x05, 0x5b, 0x17, 0x03, 0x38, 0x00, 0xf0, 0xff, 0x08, 0x08, 0x4c, 0x00, 0x4e, 0xff, 0x4c, 0x00, 0xdc, 0xff,
+    0x08, 0x5b, 0xb7, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x10, 0x00, 0x15, 0x07, 0x20, 0x00, 0x4c, 0x00, 0x36, 0xff, 0x0b, 0x0b, 0x5c, 0x5b, 0x1b, 0x07, 0x20, 0x00, 0x09, 0x00, 0xf8, 0xff,
+    0x01, 0x00, 0x4c, 0x00, 0xe8, 0xff, 0x5c, 0x5b, 0x1a, 0x07, 0x10, 0x00, 0xec, 0xff, 0x00, 0x00, 0x08, 0x08, 0x4c, 0x00, 0xce, 0xff, 0x08, 0x5b, 0xb7, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+    0x10, 0x00, 0xb7, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0xa0, 0x00, 0xb7, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x90, 0x01, 0x1a, 0x03, 0x10, 0x00, 0x00, 0x00, 0x0a, 0x00, 0x4c, 0x00,
+    0xec, 0xff, 0x40, 0x36, 0x5c, 0x5b, 0x12, 0x00, 0x08, 0xff, 0x21, 0x03, 0x00, 0x00, 0x19, 0x00, 0x00, 0x00, 0x01, 0x00, 0xff, 0xff, 0xff, 0xff, 0x00, 0x00, 0x4c, 0x00, 0xda, 0xff, 0x5c, 0x5b,
+    0x1a, 0x03, 0x10, 0x00, 0x00, 0x00, 0x0a, 0x00, 0x4c, 0x00, 0xb8, 0xff, 0x40, 0x36, 0x5c, 0x5b, 0x12, 0x00, 0xd8, 0xff, 0x1a, 0x03, 0x18, 0x00, 0x00, 0x00, 0x00, 0x00, 0x08, 0x40, 0x4c, 0x00,
+    0xe0, 0xff, 0x5c, 0x5b, 0x21, 0x03, 0x00, 0x00, 0x19, 0x00, 0x00, 0x00, 0x01, 0x00, 0xff, 0xff, 0xff, 0xff, 0x00, 0x00, 0x4c, 0x00, 0xde, 0xff, 0x5c, 0x5b, 0x1a, 0x03, 0x10, 0x00, 0x00, 0x00,
+    0x0a, 0x00, 0x4c, 0x00, 0x74, 0xff, 0x40, 0x36, 0x5c, 0x5b, 0x12, 0x00, 0xd8, 0xff, 0x1a, 0x03, 0x20, 0x00, 0x00, 0x00, 0x0a, 0x00, 0x36, 0x08, 0x40, 0x4c, 0x00, 0xdf, 0xff, 0x5b, 0x12, 0x00,
+    0x10, 0xff, 0xb7, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x10, 0x00, 0x15, 0x07, 0x28, 0x00, 0x08, 0x40, 0x0b, 0x4c, 0x00, 0x53, 0xfe, 0x0b, 0x5c, 0x5b, 0x1b, 0x07, 0x28, 0x00, 0x09, 0x00,
+    0xf8, 0xff, 0x01, 0x00, 0x4c, 0x00, 0xe6, 0xff, 0x5c, 0x5b, 0x1a, 0x07, 0x08, 0x00, 0xec, 0xff, 0x00, 0x00, 0x4c, 0x00, 0xce, 0xff, 0x40, 0x5b, 0x1a, 0x03, 0x40, 0x00, 0x00, 0x00, 0x0c, 0x00,
+    0x36, 0x4c, 0x00, 0xab, 0xff, 0x08, 0x40, 0x36, 0x36, 0x5b, 0x12, 0x00, 0xec, 0xff, 0x12, 0x00, 0x18, 0xfe, 0x12, 0x00, 0xd6, 0xff, 0x15, 0x07, 0x30, 0x00, 0x0b, 0x4c, 0x00, 0xaf, 0xff, 0x5b,
+    0x1a, 0x07, 0x58, 0x00, 0x00, 0x00, 0x10, 0x00, 0x36, 0x08, 0x40, 0x4c, 0x00, 0x09, 0xff, 0x08, 0x40, 0x4c, 0x00, 0xe3, 0xff, 0x5b, 0x12, 0x00, 0x98, 0xfe, 0x21, 0x07, 0x00, 0x00, 0x19, 0x00,
+    0x94, 0x00, 0x01, 0x00, 0xff, 0xff, 0xff, 0xff, 0x00, 0x00, 0x4c, 0x00, 0xd4, 0xff, 0x5c, 0x5b, 0x1a, 0x07, 0xa8, 0x00, 0x00, 0x00, 0x28, 0x00, 0x4c, 0x00, 0xce, 0xfd, 0x4c, 0x00, 0xca, 0xfd,
+    0x36, 0x4c, 0x00, 0xd1, 0xfd, 0x4c, 0x00, 0xcd, 0xfd, 0x36, 0x4c, 0x00, 0x2a, 0xfe, 0x08, 0x08, 0x08, 0x40, 0x36, 0x08, 0x08, 0x08, 0x4c, 0x00, 0x32, 0xfe, 0x36, 0x08, 0x40, 0x5b, 0x12, 0x00,
+    0x50, 0xfe, 0x12, 0x00, 0x84, 0xfe, 0x12, 0x00, 0x70, 0xff, 0x12, 0x00, 0xae, 0xff, 0x1a, 0x03, 0x18, 0x00, 0x00, 0x00, 0x08, 0x00, 0x08, 0x40, 0x36, 0x36, 0x5c, 0x5b, 0x12, 0x08, 0x25, 0x5c,
+    0x12, 0x08, 0x25, 0x5c, 0x21, 0x03, 0x00, 0x00, 0x19, 0x00, 0x00, 0x00, 0x01, 0x00, 0xff, 0xff, 0xff, 0xff, 0x00, 0x00, 0x4c, 0x00, 0xd8, 0xff, 0x5c, 0x5b, 0x1a, 0x03, 0x10, 0x00, 0x00, 0x00,
+    0x06, 0x00, 0x08, 0x40, 0x36, 0x5b, 0x12, 0x00, 0xdc, 0xff, 0x1a, 0x03, 0x08, 0x00, 0x00, 0x00, 0x04, 0x00, 0x36, 0x5b, 0x12, 0x00, 0xe4, 0xff, 0xb7, 0x08, 0x00, 0x00, 0x00, 0x00, 0x10, 0x27,
+    0x00, 0x00, 0x1a, 0x03, 0x88, 0x00, 0x00, 0x00, 0x1e, 0x00, 0x36, 0x36, 0x36, 0x36, 0x36, 0x36, 0x36, 0x08, 0x08, 0x08, 0x4c, 0x00, 0x32, 0xfd, 0x4c, 0x00, 0x2e, 0xfd, 0x4c, 0x00, 0x2a, 0xfd,
+    0x4c, 0x00, 0x26, 0xfd, 0x40, 0x5b, 0x12, 0x08, 0x25, 0x5c, 0x12, 0x08, 0x25, 0x5c, 0x12, 0x08, 0x25, 0x5c, 0x12, 0x08, 0x25, 0x5c, 0x12, 0x08, 0x25, 0x5c, 0x12, 0x08, 0x25, 0x5c, 0x12, 0x08,
+    0x25, 0x5c, 0x21, 0x03, 0x00, 0x00, 0x19, 0x00, 0x00, 0x00, 0x01, 0x00, 0xff, 0xff, 0xff, 0xff, 0x00, 0x00, 0x4c, 0x00, 0xae, 0xff, 0x5c, 0x5b, 0x1a, 0x03, 0x10, 0x00, 0x00, 0x00, 0x0a, 0x00,
+    0x4c, 0x00, 0x96, 0xff, 0x40, 0x36, 0x5c, 0x5b, 0x12, 0x00, 0xd8, 0xff, 0x12, 0x00, 0xda, 0xfc, 0x12, 0x00, 0x18, 0x00, 0xb7, 0x08, 0x01, 0x00, 0x00, 0x00, 0x10, 0x27, 0x00, 0x00, 0x1b, 0x00,
+    0x01, 0x00, 0x09, 0x00, 0xfc, 0xff, 0x01, 0x00, 0x02, 0x5b, 0x1a, 0x03, 0x04, 0x00, 0xf0, 0xff, 0x00, 0x00, 0x4c, 0x00, 0xe0, 0xff, 0x5c, 0x5b, 0x11, 0x14, 0xd6, 0xff, 0x11, 0x04, 0x02, 0x00,
+    0x30, 0xa0, 0x00, 0x00, 0x11, 0x04, 0x02, 0x00, 0x30, 0xe1, 0x00, 0x00, 0x30, 0x41, 0x00, 0x00, 0x11, 0x00, 0x02, 0x00, 0x2b, 0x09, 0x29, 0x00, 0x08, 0x00, 0x01, 0x00, 0x02, 0x00, 0x80, 0x00,
+    0x01, 0x00, 0x08, 0x00, 0x00, 0x00, 0x64, 0x00, 0xff, 0xff, 0x15, 0x07, 0x08, 0x00, 0x0b, 0x5b, 0xb7, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x10, 0x00, 0x15, 0x07, 0x18, 0x00, 0x4c, 0x00,
+    0x68, 0xfc, 0x0b, 0x5b, 0x1b, 0x07, 0x18, 0x00, 0x09, 0x00, 0xf8, 0xff, 0x01, 0x00, 0x4c, 0x00, 0xea, 0xff, 0x5c, 0x5b, 0x1a, 0x07, 0x10, 0x00, 0xec, 0xff, 0x00, 0x00, 0x08, 0x08, 0x4c, 0x00,
+    0xd0, 0xff, 0x08, 0x5b, 0xb7, 0x08, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x10, 0x00, 0x1b, 0x03, 0x04, 0x00, 0x09, 0x00, 0xfc, 0xff, 0x01, 0x00, 0x08, 0x5b, 0x1a, 0x03, 0x0c, 0x00, 0xf0, 0xff,
+    0x00, 0x00, 0x08, 0x08, 0x4c, 0x00, 0xde, 0xff, 0x5c, 0x5b, 0x1a, 0x07, 0x80, 0x00, 0x00, 0x00, 0x20, 0x00, 0x4c, 0x00, 0x14, 0xfc, 0x4c, 0x00, 0x10, 0xfc, 0x36, 0x4c, 0x00, 0x17, 0xfc, 0x36,
+    0x08, 0x08, 0x08, 0x08, 0x4c, 0x00, 0x84, 0xff, 0x36, 0x36, 0x4c, 0x00, 0x6a, 0xfc, 0x5c, 0x5b, 0x11, 0x00, 0x9e, 0xfc, 0x12, 0x00, 0x9e, 0xff, 0x12, 0x00, 0xc0, 0xff, 0x12, 0x00, 0xbc, 0xff,
+    0x11, 0x0c, 0x08, 0x5c, 0x11, 0x00, 0x02, 0x00, 0x2b, 0x09, 0x29, 0x54, 0x18, 0x00, 0x01, 0x00, 0x02, 0x00, 0xa8, 0x00, 0x01, 0x00, 0x06, 0x00, 0x00, 0x00, 0xf6, 0xfd, 0xff, 0xff, 0x11, 0x00,
+    0x02, 0x00, 0x2b, 0x09, 0x29, 0x00, 0x08, 0x00, 0x01, 0x00, 0x02, 0x00, 0x28, 0x00, 0x01, 0x00, 0x01, 0x00, 0x00, 0x00, 0x04, 0x00, 0xff, 0xff, 0x1a, 0x03, 0x28, 0x00, 0x00, 0x00, 0x0c, 0x00,
+    0x36, 0x36, 0x4c, 0x00, 0xa4, 0xfb, 0x08, 0x40, 0x5c, 0x5b, 0x11, 0x00, 0x44, 0xfc, 0x11, 0x08, 0x22, 0x5c, 0x11, 0x00, 0x02, 0x00, 0x2b, 0x09, 0x29, 0x00, 0x08, 0x00, 0x01, 0x00, 0x02, 0x00,
+    0x20, 0x00, 0x01, 0x00, 0x01, 0x00, 0x00, 0x00, 0x24, 0x00, 0xff, 0xff, 0xb7, 0x08, 0x01, 0x00, 0x00, 0x00, 0x10, 0x27, 0x00, 0x00, 0x21, 0x03, 0x00, 0x00, 0x19, 0x00, 0x14, 0x00, 0x01, 0x00,
+    0xff, 0xff, 0xff, 0xff, 0x00, 0x00, 0x12, 0x08, 0x25, 0x5c, 0x5c, 0x5b, 0x1a, 0x03, 0x20, 0x00, 0x00, 0x00, 0x0e, 0x00, 0x08, 0x08, 0x08, 0x08, 0x08, 0x4c, 0x00, 0xd1, 0xff, 0x36, 0x5c, 0x5b,
+    0x12, 0x00, 0xd4, 0xff, 0x11, 0x04, 0x02, 0x00, 0x2b, 0x09, 0x29, 0x54, 0x18, 0x00, 0x01, 0x00, 0x02, 0x00, 0x08, 0x00, 0x01, 0x00, 0x01, 0x00, 0x00, 0x00, 0xd0, 0xfd, 0xff, 0xff, 0x11, 0x00,
+    0x02, 0x00, 0x2b, 0x09, 0x29, 0x00, 0x08, 0x00, 0x01, 0x00, 0x02, 0x00, 0x10, 0x00, 0x01, 0x00, 0x01, 0x00, 0x00, 0x00, 0x04, 0x00, 0xff, 0xff, 0x1a, 0x03, 0x10, 0x00, 0x00, 0x00, 0x06, 0x00,
+    0x36, 0x08, 0x40, 0x5b, 0x12, 0x08, 0x25, 0x5c, 0x11, 0x04, 0x02, 0x00, 0x2b, 0x09, 0x29, 0x54, 0x18, 0x00, 0x01, 0x00, 0x02, 0x00, 0x10, 0x00, 0x01, 0x00, 0x02, 0x00, 0x00, 0x00, 0xfa, 0xfd,
+    0xff, 0xff, 0x00
 }};
 #elif defined _M_IX86
-typedef struct _ms2Ddrsr_MIDL_TYPE_FORMAT_STRING {
-	SHORT Pad;
-	UCHAR Format[2411];
-} ms2Ddrsr_MIDL_TYPE_FORMAT_STRING;
-
-typedef struct _ms2Ddrsr_MIDL_PROC_FORMAT_STRING {
-	SHORT Pad;
-	UCHAR Format[299];
-} ms2Ddrsr_MIDL_PROC_FORMAT_STRING;
-
-extern const ms2Ddrsr_MIDL_TYPE_FORMAT_STRING ms2Ddrsr__MIDL_TypeFormatString;
-extern const ms2Ddrsr_MIDL_PROC_FORMAT_STRING ms2Ddrsr__MIDL_ProcFormatString;
-static const RPC_CLIENT_INTERFACE drsuapi___RpcClientInterface = {sizeof(RPC_CLIENT_INTERFACE), {{0xe3514235, 0x4b06, 0x11d1, {0xab, 0x04, 0x00, 0xc0, 0x4f, 0xc2, 0xdc, 0xd2}}, {4, 0}}, {{0x8a885d04, 0x1ceb, 0x11c9, {0x9f, 0xe8, 0x08, 0x00, 0x2b, 0x10, 0x48, 0x60}}, {2, 0}}, 0, 0, 0, 0, 0, 0x00000000};
-static RPC_BINDING_HANDLE drsuapi__MIDL_AutoBindHandle;
-static const MIDL_STUB_DESC drsuapi_StubDesc = {(void *) &drsuapi___RpcClientInterface, MIDL_user_allocate, MIDL_user_free, &drsuapi__MIDL_AutoBindHandle, 0, 0, 0, 0, ms2Ddrsr__MIDL_TypeFormatString.Format, 1, 0x60000, 0, 0x8000253, 0, 0, 0, 0x1, 0, 0, 0};
-static const MIDL_TYPE_PICKLING_INFO __MIDL_TypePicklingInfo = {0x33205054, 0x3, 0, 0, 0,};
 #pragma optimize("", off)
 ULONG IDL_DRSBind(handle_t rpc_handle, UUID *puuidClientDsa, DRS_EXTENSIONS *pextClient, DRS_EXTENSIONS **ppextServer, DRS_HANDLE *phDrs)
 {
-	return NdrClientCall2((PMIDL_STUB_DESC) &drsuapi_StubDesc, (PFORMAT_STRING) &ms2Ddrsr__MIDL_ProcFormatString.Format[0], (unsigned char *) &rpc_handle).Simple;
+	return NdrClientCall2((PMIDL_STUB_DESC) &drsuapi_c_StubDesc, (PFORMAT_STRING) &ms2Ddrsr__MIDL_ProcFormatString.Format[0], (unsigned char *) &rpc_handle).Simple;
 }
 
 ULONG IDL_DRSUnbind(DRS_HANDLE *phDrs)
 {
-	return NdrClientCall2((PMIDL_STUB_DESC) &drsuapi_StubDesc, (PFORMAT_STRING) &ms2Ddrsr__MIDL_ProcFormatString.Format[58], (unsigned char *) &phDrs).Simple;
+	return NdrClientCall2((PMIDL_STUB_DESC) &drsuapi_c_StubDesc, (PFORMAT_STRING) &ms2Ddrsr__MIDL_ProcFormatString.Format[58], (unsigned char *) &phDrs).Simple;
 }
 
 ULONG IDL_DRSGetNCChanges(DRS_HANDLE hDrs, DWORD dwInVersion, DRS_MSG_GETCHGREQ *pmsgIn, DWORD *pdwOutVersion, DRS_MSG_GETCHGREPLY *pmsgOut)
 {
-	return NdrClientCall2((PMIDL_STUB_DESC) &drsuapi_StubDesc, (PFORMAT_STRING) &ms2Ddrsr__MIDL_ProcFormatString.Format[100], (unsigned char *) &hDrs).Simple;
+	return NdrClientCall2((PMIDL_STUB_DESC) &drsuapi_c_StubDesc, (PFORMAT_STRING) &ms2Ddrsr__MIDL_ProcFormatString.Format[128], (unsigned char *) &hDrs).Simple;
 }
 
 ULONG IDL_DRSCrackNames(DRS_HANDLE hDrs, DWORD dwInVersion, DRS_MSG_CRACKREQ *pmsgIn, DWORD *pdwOutVersion, DRS_MSG_CRACKREPLY *pmsgOut)
 {
-	return NdrClientCall2((PMIDL_STUB_DESC) &drsuapi_StubDesc, (PFORMAT_STRING) &ms2Ddrsr__MIDL_ProcFormatString.Format[166], (unsigned char *) &hDrs).Simple;
+	return NdrClientCall2((PMIDL_STUB_DESC) &drsuapi_c_StubDesc, (PFORMAT_STRING) &ms2Ddrsr__MIDL_ProcFormatString.Format[444], (unsigned char *) &hDrs).Simple;
 }
 
 ULONG IDL_DRSDomainControllerInfo(DRS_HANDLE hDrs, DWORD dwInVersion, DRS_MSG_DCINFOREQ *pmsgIn, DWORD *pdwOutVersion, DRS_MSG_DCINFOREPLY *pmsgOut)
 {
-	return NdrClientCall2((PMIDL_STUB_DESC) &drsuapi_StubDesc, (PFORMAT_STRING) &ms2Ddrsr__MIDL_ProcFormatString.Format[232], (unsigned char *) &hDrs).Simple;
+	return NdrClientCall2((PMIDL_STUB_DESC) &drsuapi_c_StubDesc, (PFORMAT_STRING) &ms2Ddrsr__MIDL_ProcFormatString.Format[594], (unsigned char *) &hDrs).Simple;
 }
 #pragma optimize("", on)
 void DRS_MSG_GETCHGREPLY_V6_Free(handle_t _MidlEsHandle, DRS_MSG_GETCHGREPLY_V6 * _pType)
 {
-    NdrMesTypeFree2(_MidlEsHandle, (PMIDL_TYPE_PICKLING_INFO) &__MIDL_TypePicklingInfo, &drsuapi_StubDesc, (PFORMAT_STRING) &ms2Ddrsr__MIDL_TypeFormatString.Format[546], _pType);
+    NdrMesTypeFree2(_MidlEsHandle, (PMIDL_TYPE_PICKLING_INFO) &__MIDL_TypePicklingInfo, &drsuapi_c_StubDesc, (PFORMAT_STRING) &ms2Ddrsr__MIDL_TypeFormatString.Format[546], _pType);
 }
 
 void DRS_MSG_CRACKREPLY_V1_Free(handle_t _MidlEsHandle, DRS_MSG_CRACKREPLY_V1 * _pType)
 {
-    NdrMesTypeFree2(_MidlEsHandle, (PMIDL_TYPE_PICKLING_INFO ) &__MIDL_TypePicklingInfo, &drsuapi_StubDesc, (PFORMAT_STRING) &ms2Ddrsr__MIDL_TypeFormatString.Format[702], _pType);
+    NdrMesTypeFree2(_MidlEsHandle, (PMIDL_TYPE_PICKLING_INFO ) &__MIDL_TypePicklingInfo, &drsuapi_c_StubDesc, (PFORMAT_STRING) &ms2Ddrsr__MIDL_TypeFormatString.Format[702], _pType);
 }
 
 void DRS_MSG_DCINFOREPLY_V2_Free(handle_t _MidlEsHandle, DRS_MSG_DCINFOREPLY_V2 * _pType)
 {
-    NdrMesTypeFree2(_MidlEsHandle, (PMIDL_TYPE_PICKLING_INFO  )&__MIDL_TypePicklingInfo, &drsuapi_StubDesc, (PFORMAT_STRING) &ms2Ddrsr__MIDL_TypeFormatString.Format[918], _pType);
+    NdrMesTypeFree2(_MidlEsHandle, (PMIDL_TYPE_PICKLING_INFO  )&__MIDL_TypePicklingInfo, &drsuapi_c_StubDesc, (PFORMAT_STRING) &ms2Ddrsr__MIDL_TypeFormatString.Format[918], _pType);
 }
 
 static const ms2Ddrsr_MIDL_PROC_FORMAT_STRING ms2Ddrsr__MIDL_ProcFormatString = {0, {
 	0x00, 0x48, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x18, 0x00, 0x32, 0x00, 0x00, 0x00, 0x44, 0x00, 0x40, 0x00, 0x47, 0x05, 0x08, 0x07, 0x01, 0x00, 0x01, 0x00, 0x00, 0x00, 0x0a, 0x00, 0x04, 0x00,
-	0x60, 0x02, 0x0b, 0x00, 0x08, 0x00, 0x64, 0x02, 0x13, 0x20, 0x0c, 0x00, 0x8c, 0x02, 0x10, 0x01, 0x10, 0x00, 0x94, 0x02, 0x70, 0x00, 0x14, 0x00, 0x08, 0x00, 0x00, 0x48, 0x00, 0x00, 0x00, 0x00,
-	0x01, 0x00, 0x08, 0x00, 0x30, 0xe0, 0x00, 0x00, 0x00, 0x00, 0x38, 0x00, 0x40, 0x00, 0x44, 0x02, 0x08, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x18, 0x01, 0x00, 0x00, 0x9c, 0x02, 0x70, 0x00,
-	0x04, 0x00, 0x08, 0x00, 0x00, 0x48, 0x00, 0x00, 0x00, 0x00, 0x03, 0x00, 0x18, 0x00, 0x30, 0x40, 0x00, 0x00, 0x00, 0x00, 0x2c, 0x00, 0x24, 0x00, 0x47, 0x06, 0x08, 0x07, 0x01, 0x00, 0x01, 0x00,
-	0x00, 0x00, 0x08, 0x00, 0x00, 0x00, 0xa0, 0x02, 0x48, 0x00, 0x04, 0x00, 0x08, 0x00, 0x0b, 0x01, 0x08, 0x00, 0xa8, 0x02, 0x50, 0x21, 0x0c, 0x00, 0x08, 0x00, 0x13, 0x01, 0x10, 0x00, 0x68, 0x04,
-	0x70, 0x00, 0x14, 0x00, 0x08, 0x00, 0x00, 0x48, 0x00, 0x00, 0x00, 0x00, 0x0c, 0x00, 0x18, 0x00, 0x30, 0x40, 0x00, 0x00, 0x00, 0x00, 0x2c, 0x00, 0x24, 0x00, 0x47, 0x06, 0x08, 0x07, 0x01, 0x00,
-	0x01, 0x00, 0x00, 0x00, 0x08, 0x00, 0x00, 0x00, 0xa0, 0x02, 0x48, 0x00, 0x04, 0x00, 0x08, 0x00, 0x0b, 0x01, 0x08, 0x00, 0x86, 0x05, 0x50, 0x21, 0x0c, 0x00, 0x08, 0x00, 0x13, 0x21, 0x10, 0x00,
-	0xe2, 0x05, 0x70, 0x00, 0x14, 0x00, 0x08, 0x00, 0x00, 0x48, 0x00, 0x00, 0x00, 0x00, 0x10, 0x00, 0x18, 0x00, 0x30, 0x40, 0x00, 0x00, 0x00, 0x00, 0x2c, 0x00, 0x24, 0x00, 0x47, 0x06, 0x08, 0x07,
-	0x01, 0x00, 0x01, 0x00, 0x00, 0x00, 0x08, 0x00, 0x00, 0x00, 0xa0, 0x02, 0x48, 0x00, 0x04, 0x00, 0x08, 0x00, 0x0b, 0x01, 0x08, 0x00, 0x6e, 0x06, 0x50, 0x21, 0x0c, 0x00, 0x08, 0x00, 0x13, 0x21,
-	0x10, 0x00, 0x9c, 0x06, 0x70, 0x00, 0x14, 0x00, 0x08, 0x00, 0x00,
+	0xa8, 0x03, 0x0b, 0x00, 0x08, 0x00, 0xac, 0x03, 0x13, 0x20, 0x0c, 0x00, 0xd4, 0x03, 0x10, 0x01, 0x10, 0x00, 0xdc, 0x03, 0x70, 0x00, 0x14, 0x00, 0x08, 0x00, 0x00, 0x48, 0x00, 0x00, 0x00, 0x00,
+	0x01, 0x00, 0x08, 0x00, 0x30, 0xe0, 0x00, 0x00, 0x00, 0x00, 0x38, 0x00, 0x40, 0x00, 0x44, 0x02, 0x08, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x18, 0x01, 0x00, 0x00, 0xe4, 0x03, 0x70, 0x00,
+	0x04, 0x00, 0x08, 0x00, 0x00, 0x48, 0x00, 0x00, 0x00, 0x00, 0x02, 0x00, 0x04, 0x00, 0x32, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x40, 0x00, 0x08, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+	0x00, 0x48, 0x00, 0x00, 0x00, 0x00, 0x03, 0x00, 0x18, 0x00, 0x30, 0x40, 0x00, 0x00, 0x00, 0x00, 0x2c, 0x00, 0x24, 0x00, 0x47, 0x06, 0x08, 0x07, 0x01, 0x00, 0x01, 0x00, 0x00, 0x00, 0x08, 0x00,
+	0x00, 0x00, 0xe8, 0x03, 0x48, 0x00, 0x04, 0x00, 0x08, 0x00, 0x0b, 0x01, 0x08, 0x00, 0xf0, 0x03, 0x50, 0x21, 0x0c, 0x00, 0x08, 0x00, 0x13, 0x01, 0x10, 0x00, 0xa6, 0x04, 0x70, 0x00, 0x14, 0x00,
+	0x08, 0x00, 0x00, 0x48, 0x00, 0x00, 0x00, 0x00, 0x04, 0x00, 0x10, 0x00, 0x30, 0x40, 0x00, 0x00, 0x00, 0x00, 0x2c, 0x00, 0x08, 0x00, 0x46, 0x04, 0x08, 0x05, 0x00, 0x00, 0x01, 0x00, 0x00, 0x00,
+	0x08, 0x00, 0x00, 0x00, 0xe8, 0x03, 0x48, 0x00, 0x04, 0x00, 0x08, 0x00, 0x0b, 0x01, 0x08, 0x00, 0xc0, 0x04, 0x70, 0x00, 0x0c, 0x00, 0x08, 0x00, 0x00, 0x48, 0x00, 0x00, 0x00, 0x00, 0x05, 0x00,
+	0x04, 0x00, 0x32, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x40, 0x00, 0x08, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x48, 0x00, 0x00, 0x00, 0x00, 0x06, 0x00, 0x04, 0x00, 0x32, 0x00,
+	0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x40, 0x00, 0x08, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x48, 0x00, 0x00, 0x00, 0x00, 0x07, 0x00, 0x04, 0x00, 0x32, 0x00, 0x00, 0x00, 0x00, 0x00,
+	0x00, 0x00, 0x40, 0x00, 0x08, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x48, 0x00, 0x00, 0x00, 0x00, 0x08, 0x00, 0x04, 0x00, 0x32, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x40, 0x00,
+	0x08, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x48, 0x00, 0x00, 0x00, 0x00, 0x09, 0x00, 0x04, 0x00, 0x32, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x40, 0x00, 0x08, 0x01, 0x00, 0x00,
+	0x00, 0x00, 0x00, 0x00, 0x00, 0x48, 0x00, 0x00, 0x00, 0x00, 0x0a, 0x00, 0x04, 0x00, 0x32, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x40, 0x00, 0x08, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+	0x00, 0x48, 0x00, 0x00, 0x00, 0x00, 0x0b, 0x00, 0x04, 0x00, 0x32, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x40, 0x00, 0x08, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x48, 0x00, 0x00,
+	0x00, 0x00, 0x0c, 0x00, 0x18, 0x00, 0x30, 0x40, 0x00, 0x00, 0x00, 0x00, 0x2c, 0x00, 0x24, 0x00, 0x47, 0x06, 0x08, 0x07, 0x01, 0x00, 0x01, 0x00, 0x00, 0x00, 0x08, 0x00, 0x00, 0x00, 0xe8, 0x03,
+	0x48, 0x00, 0x04, 0x00, 0x08, 0x00, 0x0b, 0x01, 0x08, 0x00, 0xfe, 0x04, 0x50, 0x21, 0x0c, 0x00, 0x08, 0x00, 0x13, 0x21, 0x10, 0x00, 0x5a, 0x05, 0x70, 0x00, 0x14, 0x00, 0x08, 0x00, 0x00, 0x48,
+	0x00, 0x00, 0x00, 0x00, 0x0d, 0x00, 0x04, 0x00, 0x32, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x40, 0x00, 0x08, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x48, 0x00, 0x00, 0x00, 0x00,
+	0x0e, 0x00, 0x04, 0x00, 0x32, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x40, 0x00, 0x08, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x48, 0x00, 0x00, 0x00, 0x00, 0x0f, 0x00, 0x04, 0x00,
+	0x32, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x40, 0x00, 0x08, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x48, 0x00, 0x00, 0x00, 0x00, 0x10, 0x00, 0x18, 0x00, 0x30, 0x40, 0x00, 0x00,
+	0x00, 0x00, 0x2c, 0x00, 0x24, 0x00, 0x47, 0x06, 0x08, 0x07, 0x01, 0x00, 0x01, 0x00, 0x00, 0x00, 0x08, 0x00, 0x00, 0x00, 0xe8, 0x03, 0x48, 0x00, 0x04, 0x00, 0x08, 0x00, 0x0b, 0x01, 0x08, 0x00,
+	0x74, 0x05, 0x50, 0x21, 0x0c, 0x00, 0x08, 0x00, 0x13, 0x21, 0x10, 0x00, 0xa2, 0x05, 0x70, 0x00, 0x14, 0x00, 0x08, 0x00, 0x00
 }};
 static const ms2Ddrsr_MIDL_TYPE_FORMAT_STRING ms2Ddrsr__MIDL_TypeFormatString = {0, {
 	0x00, 0x00, 0x1d, 0x00, 0x08, 0x00, 0x01, 0x5b, 0x15, 0x03, 0x10, 0x00, 0x08, 0x06, 0x06, 0x4c, 0x00, 0xf1, 0xff, 0x5b, 0x15, 0x07, 0x18, 0x00, 0x0b, 0x0b, 0x0b, 0x5b, 0xb7, 0x08, 0x00, 0x00,
@@ -227,62 +256,46 @@ static const ms2Ddrsr_MIDL_TYPE_FORMAT_STRING ms2Ddrsr__MIDL_TypeFormatString = 
 	0x08, 0x40, 0x4c, 0x00, 0xe4, 0xff, 0x5c, 0x5b, 0x12, 0x00, 0xa0, 0xfe, 0x21, 0x07, 0x00, 0x00, 0x19, 0x00, 0x80, 0x00, 0x01, 0x00, 0xff, 0xff, 0xff, 0xff, 0x00, 0x00, 0x4c, 0x00, 0xd4, 0xff,
 	0x5c, 0x5b, 0x1a, 0x07, 0x90, 0x00, 0x00, 0x00, 0x28, 0x00, 0x4c, 0x00, 0xdc, 0xfd, 0x4c, 0x00, 0xd8, 0xfd, 0x36, 0x40, 0x4c, 0x00, 0xde, 0xfd, 0x4c, 0x00, 0xda, 0xfd, 0x36, 0x4c, 0x00, 0x33,
 	0xfe, 0x08, 0x08, 0x08, 0x36, 0x08, 0x08, 0x08, 0x4c, 0x00, 0x3a, 0xfe, 0x36, 0x08, 0x40, 0x5b, 0x12, 0x00, 0x58, 0xfe, 0x12, 0x00, 0x8c, 0xfe, 0x12, 0x00, 0x70, 0xff, 0x12, 0x00, 0xae, 0xff,
-	0x12, 0x00, 0xa6, 0xfd, 0x12, 0x00, 0x18, 0x00, 0xb7, 0x08, 0x01, 0x00, 0x00, 0x00, 0x10, 0x27, 0x00, 0x00, 0x1b, 0x00, 0x01, 0x00, 0x09, 0x00, 0xfc, 0xff, 0x01, 0x00, 0x02, 0x5b, 0x1a, 0x03,
-	0x04, 0x00, 0xf0, 0xff, 0x00, 0x00, 0x4c, 0x00, 0xe0, 0xff, 0x5c, 0x5b, 0x11, 0x14, 0xd6, 0xff, 0x11, 0x04, 0x02, 0x00, 0x30, 0xa0, 0x00, 0x00, 0x11, 0x04, 0x02, 0x00, 0x30, 0xe1, 0x00, 0x00,
-	0x30, 0x41, 0x00, 0x00, 0x11, 0x00, 0x02, 0x00, 0x2b, 0x09, 0x29, 0x00, 0x04, 0x00, 0x01, 0x00, 0x02, 0x00, 0x88, 0x00, 0x05, 0x00, 0x04, 0x00, 0x00, 0x00, 0xbc, 0x00, 0x05, 0x00, 0x00, 0x00,
-	0xd4, 0x00, 0x07, 0x00, 0x00, 0x00, 0x02, 0x01, 0x08, 0x00, 0x00, 0x00, 0x22, 0x01, 0x0a, 0x00, 0x00, 0x00, 0x54, 0x01, 0xff, 0xff, 0xb7, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x10, 0x00,
-	0x15, 0x07, 0x18, 0x00, 0x4c, 0x00, 0x22, 0xfd, 0x0b, 0x5b, 0x1b, 0x07, 0x18, 0x00, 0x09, 0x00, 0xf8, 0xff, 0x01, 0x00, 0x4c, 0x00, 0xea, 0xff, 0x5c, 0x5b, 0x1a, 0x07, 0x10, 0x00, 0xec, 0xff,
-	0x00, 0x00, 0x08, 0x08, 0x4c, 0x00, 0xd0, 0xff, 0x08, 0x5b, 0xb7, 0x08, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x10, 0x00, 0x1b, 0x03, 0x04, 0x00, 0x09, 0x00, 0xfc, 0xff, 0x01, 0x00, 0x08, 0x5b,
-	0x1a, 0x03, 0x0c, 0x00, 0xf0, 0xff, 0x00, 0x00, 0x08, 0x08, 0x4c, 0x00, 0xde, 0xff, 0x5c, 0x5b, 0x1a, 0x07, 0x60, 0x00, 0x00, 0x00, 0x1c, 0x00, 0x4c, 0x00, 0xce, 0xfc, 0x4c, 0x00, 0xca, 0xfc,
-	0x36, 0x40, 0x4c, 0x00, 0xd0, 0xfc, 0x36, 0x36, 0x4c, 0x00, 0x28, 0xfd, 0x08, 0x08, 0x08, 0x08, 0x5c, 0x5b, 0x11, 0x00, 0x56, 0xfd, 0x12, 0x00, 0xa2, 0xff, 0x12, 0x00, 0xc4, 0xff, 0xb7, 0x08,
-	0x01, 0x00, 0x00, 0x00, 0x00, 0x01, 0x00, 0x00, 0x1a, 0x03, 0x04, 0x00, 0x06, 0xff, 0x00, 0x00, 0x4c, 0x00, 0xec, 0xff, 0x5c, 0x5b, 0x1a, 0x07, 0x78, 0x00, 0x00, 0x00, 0x0e, 0x00, 0x4c, 0x00,
-	0x88, 0xfc, 0x36, 0x40, 0x4c, 0x00, 0xaa, 0xff, 0x5c, 0x5b, 0x11, 0x00, 0xdc, 0xff, 0x15, 0x07, 0x08, 0x00, 0x0b, 0x5b, 0x16, 0x07, 0x60, 0x00, 0x4b, 0x5c, 0x46, 0x5c, 0x20, 0x00, 0x20, 0x00,
-	0x11, 0x00, 0x08, 0xfd, 0x46, 0x5c, 0x40, 0x00, 0x40, 0x00, 0x12, 0x00, 0x4e, 0xff, 0x5b, 0x4c, 0x00, 0x57, 0xfc, 0x4c, 0x00, 0x53, 0xfc, 0x08, 0x40, 0x4c, 0x00, 0x59, 0xfc, 0x08, 0x08, 0x08,
-	0x08, 0x08, 0x40, 0x4c, 0x00, 0xc9, 0xff, 0x5b, 0x1a, 0x07, 0x88, 0x00, 0x00, 0x00, 0x14, 0x00, 0x4c, 0x00, 0x36, 0xfc, 0x36, 0x40, 0x4c, 0x00, 0x58, 0xff, 0x36, 0x36, 0x4c, 0x00, 0x94, 0xfc,
-	0x5c, 0x5b, 0x11, 0x00, 0x84, 0xff, 0x12, 0x00, 0x38, 0xff, 0x12, 0x00, 0x34, 0xff, 0x1a, 0x07, 0x70, 0x00, 0x00, 0x00, 0x22, 0x00, 0x4c, 0x00, 0x10, 0xfc, 0x4c, 0x00, 0x0c, 0xfc, 0x36, 0x40,
-	0x4c, 0x00, 0x12, 0xfc, 0x36, 0x08, 0x08, 0x08, 0x08, 0x40, 0x4c, 0x00, 0x82, 0xff, 0x36, 0x36, 0x4c, 0x00, 0x60, 0xfc, 0x5c, 0x5b, 0x11, 0x00, 0x92, 0xfc, 0x12, 0x00, 0xde, 0xfe, 0x12, 0x00,
-	0x00, 0xff, 0x12, 0x00, 0xfc, 0xfe, 0x1a, 0x07, 0x78, 0x00, 0x00, 0x00, 0x24, 0x00, 0x4c, 0x00, 0xd8, 0xfb, 0x4c, 0x00, 0xd4, 0xfb, 0x36, 0x40, 0x4c, 0x00, 0xda, 0xfb, 0x36, 0x08, 0x08, 0x08,
-	0x08, 0x40, 0x4c, 0x00, 0x4a, 0xff, 0x36, 0x36, 0x4c, 0x00, 0x28, 0xfc, 0x08, 0x40, 0x5c, 0x5b, 0x11, 0x00, 0x58, 0xfc, 0x12, 0x00, 0xa4, 0xfe, 0x12, 0x00, 0xc6, 0xfe, 0x12, 0x00, 0xc2, 0xfe,
-	0x11, 0x0c, 0x08, 0x5c, 0x11, 0x00, 0x02, 0x00, 0x2b, 0x09, 0x29, 0x54, 0x0c, 0x00, 0x01, 0x00, 0x02, 0x00, 0x90, 0x00, 0x05, 0x00, 0x01, 0x00, 0x00, 0x00, 0x1c, 0x00, 0x02, 0x00, 0x00, 0x00,
-	0x54, 0x00, 0x06, 0x00, 0x00, 0x00, 0x9c, 0xfd, 0x07, 0x00, 0x00, 0x00, 0x5e, 0x00, 0x09, 0x00, 0x00, 0x00, 0xb2, 0x00, 0xff, 0xff, 0x1a, 0x07, 0x78, 0x00, 0x00, 0x00, 0x20, 0x00, 0x4c, 0x00,
-	0x68, 0xfb, 0x4c, 0x00, 0x64, 0xfb, 0x36, 0x40, 0x4c, 0x00, 0x6a, 0xfb, 0x4c, 0x00, 0x66, 0xfb, 0x36, 0x4c, 0x00, 0xbf, 0xfb, 0x08, 0x08, 0x08, 0x36, 0x08, 0x5c, 0x5b, 0x12, 0x00, 0xec, 0xfb,
-	0x12, 0x00, 0x38, 0xfe, 0x12, 0x00, 0x04, 0xfd, 0x1b, 0x00, 0x01, 0x00, 0x19, 0x00, 0x04, 0x00, 0x01, 0x00, 0x02, 0x5b, 0x16, 0x03, 0x0c, 0x00, 0x4b, 0x5c, 0x46, 0x5c, 0x08, 0x00, 0x08, 0x00,
-	0x12, 0x00, 0xe6, 0xff, 0x5b, 0x08, 0x08, 0x08, 0x5c, 0x5b, 0x1a, 0x03, 0x14, 0x00, 0x00, 0x00, 0x00, 0x00, 0x08, 0x0d, 0x4c, 0x00, 0xde, 0xff, 0x5c, 0x5b, 0xb7, 0x08, 0x00, 0x00, 0x00, 0x00,
-	0x00, 0x00, 0x10, 0x00, 0x15, 0x07, 0x48, 0x00, 0x0b, 0x4c, 0x00, 0x93, 0xfc, 0x08, 0x08, 0x08, 0x40, 0x0b, 0x5c, 0x5b, 0x1a, 0x07, 0x60, 0x00, 0x00, 0x00, 0x10, 0x00, 0x36, 0x08, 0x4c, 0x00,
-	0xf0, 0xfb, 0x08, 0x40, 0x4c, 0x00, 0xde, 0xff, 0x5c, 0x5b, 0x12, 0x00, 0x7e, 0xfb, 0x21, 0x07, 0x00, 0x00, 0x19, 0x00, 0x80, 0x00, 0x01, 0x00, 0xff, 0xff, 0xff, 0xff, 0x00, 0x00, 0x4c, 0x00,
-	0xd4, 0xff, 0x5c, 0x5b, 0x1a, 0x07, 0x90, 0x00, 0x00, 0x00, 0x28, 0x00, 0x4c, 0x00, 0xba, 0xfa, 0x4c, 0x00, 0xb6, 0xfa, 0x36, 0x40, 0x4c, 0x00, 0xbc, 0xfa, 0x4c, 0x00, 0xb8, 0xfa, 0x36, 0x4c,
-	0x00, 0x11, 0xfb, 0x08, 0x08, 0x08, 0x36, 0x08, 0x08, 0x08, 0x4c, 0x00, 0x8e, 0xff, 0x36, 0x08, 0x40, 0x5b, 0x12, 0x00, 0x36, 0xfb, 0x12, 0x00, 0x6a, 0xfb, 0x12, 0x00, 0x4e, 0xfc, 0x12, 0x00,
-	0xae, 0xff, 0x11, 0x00, 0x02, 0x00, 0x2b, 0x09, 0x29, 0x00, 0x04, 0x00, 0x01, 0x00, 0x02, 0x00, 0x1c, 0x00, 0x01, 0x00, 0x01, 0x00, 0x00, 0x00, 0x2e, 0x00, 0xff, 0xff, 0xb7, 0x08, 0x01, 0x00,
-	0x00, 0x00, 0x10, 0x27, 0x00, 0x00, 0x1b, 0x03, 0x04, 0x00, 0x19, 0x00, 0x14, 0x00, 0x01, 0x00, 0x4b, 0x5c, 0x48, 0x49, 0x04, 0x00, 0x00, 0x00, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x12, 0x08,
-	0x25, 0x5c, 0x5b, 0x08, 0x5c, 0x5b, 0x1a, 0x03, 0x1c, 0x00, 0x00, 0x00, 0x0e, 0x00, 0x08, 0x08, 0x08, 0x08, 0x08, 0x4c, 0x00, 0xc7, 0xff, 0x36, 0x5c, 0x5b, 0x12, 0x00, 0xca, 0xff, 0x11, 0x04,
-	0x02, 0x00, 0x2b, 0x09, 0x29, 0x54, 0x0c, 0x00, 0x01, 0x00, 0x02, 0x00, 0x04, 0x00, 0x01, 0x00, 0x01, 0x00, 0x00, 0x00, 0x62, 0x00, 0xff, 0xff, 0x16, 0x03, 0x0c, 0x00, 0x4b, 0x5c, 0x46, 0x5c,
-	0x04, 0x00, 0x04, 0x00, 0x12, 0x08, 0x25, 0x5c, 0x46, 0x5c, 0x08, 0x00, 0x08, 0x00, 0x12, 0x08, 0x25, 0x5c, 0x5b, 0x08, 0x08, 0x08, 0x5c, 0x5b, 0x1b, 0x03, 0x0c, 0x00, 0x19, 0x00, 0x00, 0x00,
-	0x01, 0x00, 0x4b, 0x5c, 0x48, 0x49, 0x0c, 0x00, 0x00, 0x00, 0x02, 0x00, 0x04, 0x00, 0x04, 0x00, 0x12, 0x08, 0x25, 0x5c, 0x08, 0x00, 0x08, 0x00, 0x12, 0x08, 0x25, 0x5c, 0x5b, 0x4c, 0x00, 0xb9,
-	0xff, 0x5b, 0x16, 0x03, 0x08, 0x00, 0x4b, 0x5c, 0x46, 0x5c, 0x04, 0x00, 0x04, 0x00, 0x12, 0x00, 0xc8, 0xff, 0x5b, 0x08, 0x08, 0x5b, 0x16, 0x03, 0x04, 0x00, 0x4b, 0x5c, 0x46, 0x5c, 0x00, 0x00,
-	0x00, 0x00, 0x12, 0x00, 0xde, 0xff, 0x5b, 0x08, 0x5c, 0x5b, 0x11, 0x00, 0x02, 0x00, 0x2b, 0x09, 0x29, 0x00, 0x04, 0x00, 0x01, 0x00, 0x02, 0x00, 0x08, 0x00, 0x01, 0x00, 0x01, 0x00, 0x00, 0x00,
-	0x04, 0x00, 0xff, 0xff, 0x16, 0x03, 0x08, 0x00, 0x4b, 0x5c, 0x46, 0x5c, 0x00, 0x00, 0x00, 0x00, 0x12, 0x08, 0x25, 0x5c, 0x5b, 0x08, 0x08, 0x5b, 0x11, 0x04, 0x02, 0x00, 0x2b, 0x09, 0x29, 0x54,
-	0x0c, 0x00, 0x01, 0x00, 0x02, 0x00, 0x08, 0x00, 0x04, 0x00, 0x01, 0x00, 0x00, 0x00, 0xa4, 0x00, 0x02, 0x00, 0x00, 0x00, 0x74, 0x01, 0x03, 0x00, 0x00, 0x00, 0x46, 0x02, 0xff, 0xff, 0xff, 0xff,
-	0x98, 0x02, 0xff, 0xff, 0xb7, 0x08, 0x00, 0x00, 0x00, 0x00, 0x10, 0x27, 0x00, 0x00, 0x16, 0x03, 0x1c, 0x00, 0x4b, 0x5c, 0x46, 0x5c, 0x00, 0x00, 0x00, 0x00, 0x12, 0x08, 0x25, 0x5c, 0x46, 0x5c,
-	0x04, 0x00, 0x04, 0x00, 0x12, 0x08, 0x25, 0x5c, 0x46, 0x5c, 0x08, 0x00, 0x08, 0x00, 0x12, 0x08, 0x25, 0x5c, 0x46, 0x5c, 0x0c, 0x00, 0x0c, 0x00, 0x12, 0x08, 0x25, 0x5c, 0x46, 0x5c, 0x10, 0x00,
-	0x10, 0x00, 0x12, 0x08, 0x25, 0x5c, 0x5b, 0x08, 0x08, 0x08, 0x08, 0x08, 0x08, 0x08, 0x5c, 0x5b, 0x1b, 0x03, 0x1c, 0x00, 0x19, 0x00, 0x00, 0x00, 0x01, 0x00, 0x4b, 0x5c, 0x48, 0x49, 0x1c, 0x00,
-	0x00, 0x00, 0x05, 0x00, 0x00, 0x00, 0x00, 0x00, 0x12, 0x08, 0x25, 0x5c, 0x04, 0x00, 0x04, 0x00, 0x12, 0x08, 0x25, 0x5c, 0x08, 0x00, 0x08, 0x00, 0x12, 0x08, 0x25, 0x5c, 0x0c, 0x00, 0x0c, 0x00,
-	0x12, 0x08, 0x25, 0x5c, 0x10, 0x00, 0x10, 0x00, 0x12, 0x08, 0x25, 0x5c, 0x5b, 0x4c, 0x00, 0x7f, 0xff, 0x5b, 0x1a, 0x03, 0x08, 0x00, 0x00, 0x00, 0x08, 0x00, 0x4c, 0x00, 0x68, 0xff, 0x36, 0x5b,
-	0x12, 0x00, 0xae, 0xff, 0xb7, 0x08, 0x00, 0x00, 0x00, 0x00, 0x10, 0x27, 0x00, 0x00, 0x16, 0x03, 0x68, 0x00, 0x4b, 0x5c, 0x46, 0x5c, 0x00, 0x00, 0x00, 0x00, 0x12, 0x08, 0x25, 0x5c, 0x46, 0x5c,
-	0x04, 0x00, 0x04, 0x00, 0x12, 0x08, 0x25, 0x5c, 0x46, 0x5c, 0x08, 0x00, 0x08, 0x00, 0x12, 0x08, 0x25, 0x5c, 0x46, 0x5c, 0x0c, 0x00, 0x0c, 0x00, 0x12, 0x08, 0x25, 0x5c, 0x46, 0x5c, 0x10, 0x00,
-	0x10, 0x00, 0x12, 0x08, 0x25, 0x5c, 0x46, 0x5c, 0x14, 0x00, 0x14, 0x00, 0x12, 0x08, 0x25, 0x5c, 0x46, 0x5c, 0x18, 0x00, 0x18, 0x00, 0x12, 0x08, 0x25, 0x5c, 0x5b, 0x08, 0x08, 0x08, 0x08, 0x08,
-	0x08, 0x08, 0x08, 0x08, 0x08, 0x4c, 0x00, 0x41, 0xf8, 0x4c, 0x00, 0x3d, 0xf8, 0x4c, 0x00, 0x39, 0xf8, 0x4c, 0x00, 0x35, 0xf8, 0x5b, 0x1b, 0x03, 0x68, 0x00, 0x19, 0x00, 0x00, 0x00, 0x01, 0x00,
-	0x4b, 0x5c, 0x48, 0x49, 0x68, 0x00, 0x00, 0x00, 0x07, 0x00, 0x00, 0x00, 0x00, 0x00, 0x12, 0x08, 0x25, 0x5c, 0x04, 0x00, 0x04, 0x00, 0x12, 0x08, 0x25, 0x5c, 0x08, 0x00, 0x08, 0x00, 0x12, 0x08,
-	0x25, 0x5c, 0x0c, 0x00, 0x0c, 0x00, 0x12, 0x08, 0x25, 0x5c, 0x10, 0x00, 0x10, 0x00, 0x12, 0x08, 0x25, 0x5c, 0x14, 0x00, 0x14, 0x00, 0x12, 0x08, 0x25, 0x5c, 0x18, 0x00, 0x18, 0x00, 0x12, 0x08,
-	0x25, 0x5c, 0x5b, 0x4c, 0x00, 0x49, 0xff, 0x5b, 0x1a, 0x03, 0x08, 0x00, 0x00, 0x00, 0x08, 0x00, 0x4c, 0x00, 0x32, 0xff, 0x36, 0x5b, 0x12, 0x00, 0x9e, 0xff, 0xb7, 0x08, 0x00, 0x00, 0x00, 0x00,
-	0x10, 0x27, 0x00, 0x00, 0x16, 0x03, 0x6c, 0x00, 0x4b, 0x5c, 0x46, 0x5c, 0x00, 0x00, 0x00, 0x00, 0x12, 0x08, 0x25, 0x5c, 0x46, 0x5c, 0x04, 0x00, 0x04, 0x00, 0x12, 0x08, 0x25, 0x5c, 0x46, 0x5c,
-	0x08, 0x00, 0x08, 0x00, 0x12, 0x08, 0x25, 0x5c, 0x46, 0x5c, 0x0c, 0x00, 0x0c, 0x00, 0x12, 0x08, 0x25, 0x5c, 0x46, 0x5c, 0x10, 0x00, 0x10, 0x00, 0x12, 0x08, 0x25, 0x5c, 0x46, 0x5c, 0x14, 0x00,
-	0x14, 0x00, 0x12, 0x08, 0x25, 0x5c, 0x46, 0x5c, 0x18, 0x00, 0x18, 0x00, 0x12, 0x08, 0x25, 0x5c, 0x5b, 0x08, 0x08, 0x08, 0x08, 0x08, 0x08, 0x08, 0x08, 0x08, 0x08, 0x08, 0x4c, 0x00, 0x6a, 0xf7,
-	0x4c, 0x00, 0x66, 0xf7, 0x4c, 0x00, 0x62, 0xf7, 0x4c, 0x00, 0x5e, 0xf7, 0x5c, 0x5b, 0x1b, 0x03, 0x6c, 0x00, 0x19, 0x00, 0x00, 0x00, 0x01, 0x00, 0x4b, 0x5c, 0x48, 0x49, 0x6c, 0x00, 0x00, 0x00,
-	0x07, 0x00, 0x00, 0x00, 0x00, 0x00, 0x12, 0x08, 0x25, 0x5c, 0x04, 0x00, 0x04, 0x00, 0x12, 0x08, 0x25, 0x5c, 0x08, 0x00, 0x08, 0x00, 0x12, 0x08, 0x25, 0x5c, 0x0c, 0x00, 0x0c, 0x00, 0x12, 0x08,
-	0x25, 0x5c, 0x10, 0x00, 0x10, 0x00, 0x12, 0x08, 0x25, 0x5c, 0x14, 0x00, 0x14, 0x00, 0x12, 0x08, 0x25, 0x5c, 0x18, 0x00, 0x18, 0x00, 0x12, 0x08, 0x25, 0x5c, 0x5b, 0x4c, 0x00, 0x47, 0xff, 0x5b,
-	0x1a, 0x03, 0x08, 0x00, 0x00, 0x00, 0x08, 0x00, 0x4c, 0x00, 0x30, 0xff, 0x36, 0x5b, 0x12, 0x00, 0x9e, 0xff, 0xb7, 0x08, 0x00, 0x00, 0x00, 0x00, 0x10, 0x27, 0x00, 0x00, 0x16, 0x03, 0x1c, 0x00,
-	0x4b, 0x5c, 0x46, 0x5c, 0x18, 0x00, 0x18, 0x00, 0x12, 0x08, 0x25, 0x5c, 0x5b, 0x08, 0x08, 0x08, 0x08, 0x08, 0x08, 0x08, 0x5c, 0x5b, 0x1b, 0x03, 0x1c, 0x00, 0x19, 0x00, 0x00, 0x00, 0x01, 0x00,
-	0x4b, 0x5c, 0x48, 0x49, 0x1c, 0x00, 0x00, 0x00, 0x01, 0x00, 0x18, 0x00, 0x18, 0x00, 0x12, 0x08, 0x25, 0x5c, 0x5b, 0x4c, 0x00, 0xc7, 0xff, 0x5b, 0x1a, 0x03, 0x08, 0x00, 0x00, 0x00, 0x08, 0x00,
-	0x4c, 0x00, 0xb0, 0xff, 0x36, 0x5b, 0x12, 0x00, 0xce, 0xff, 0x00,
+	0x16, 0x03, 0x0c, 0x00, 0x4b, 0x5c, 0x46, 0x5c, 0x04, 0x00, 0x04, 0x00, 0x12, 0x08, 0x25, 0x5c, 0x46, 0x5c, 0x08, 0x00, 0x08, 0x00, 0x12, 0x08, 0x25, 0x5c, 0x5b, 0x08, 0x08, 0x08, 0x5c, 0x5b,
+	0x1b, 0x03, 0x0c, 0x00, 0x19, 0x00, 0x00, 0x00, 0x01, 0x00, 0x4b, 0x5c, 0x48, 0x49, 0x0c, 0x00, 0x00, 0x00, 0x02, 0x00, 0x04, 0x00, 0x04, 0x00, 0x12, 0x08, 0x25, 0x5c, 0x08, 0x00, 0x08, 0x00,
+	0x12, 0x08, 0x25, 0x5c, 0x5b, 0x4c, 0x00, 0xb9, 0xff, 0x5b, 0x16, 0x03, 0x08, 0x00, 0x4b, 0x5c, 0x46, 0x5c, 0x04, 0x00, 0x04, 0x00, 0x12, 0x00, 0xc8, 0xff, 0x5b, 0x08, 0x08, 0x5b, 0x16, 0x03,
+	0x04, 0x00, 0x4b, 0x5c, 0x46, 0x5c, 0x00, 0x00, 0x00, 0x00, 0x12, 0x00, 0xde, 0xff, 0x5b, 0x08, 0x5c, 0x5b, 0xb7, 0x08, 0x00, 0x00, 0x00, 0x00, 0x10, 0x27, 0x00, 0x00, 0x16, 0x03, 0x68, 0x00,
+	0x4b, 0x5c, 0x46, 0x5c, 0x00, 0x00, 0x00, 0x00, 0x12, 0x08, 0x25, 0x5c, 0x46, 0x5c, 0x04, 0x00, 0x04, 0x00, 0x12, 0x08, 0x25, 0x5c, 0x46, 0x5c, 0x08, 0x00, 0x08, 0x00, 0x12, 0x08, 0x25, 0x5c,
+	0x46, 0x5c, 0x0c, 0x00, 0x0c, 0x00, 0x12, 0x08, 0x25, 0x5c, 0x46, 0x5c, 0x10, 0x00, 0x10, 0x00, 0x12, 0x08, 0x25, 0x5c, 0x46, 0x5c, 0x14, 0x00, 0x14, 0x00, 0x12, 0x08, 0x25, 0x5c, 0x46, 0x5c,
+	0x18, 0x00, 0x18, 0x00, 0x12, 0x08, 0x25, 0x5c, 0x5b, 0x08, 0x08, 0x08, 0x08, 0x08, 0x08, 0x08, 0x08, 0x08, 0x08, 0x4c, 0x00, 0xd3, 0xfc, 0x4c, 0x00, 0xcf, 0xfc, 0x4c, 0x00, 0xcb, 0xfc, 0x4c,
+	0x00, 0xc7, 0xfc, 0x5b, 0x1b, 0x03, 0x68, 0x00, 0x19, 0x00, 0x00, 0x00, 0x01, 0x00, 0x4b, 0x5c, 0x48, 0x49, 0x68, 0x00, 0x00, 0x00, 0x07, 0x00, 0x00, 0x00, 0x00, 0x00, 0x12, 0x08, 0x25, 0x5c,
+	0x04, 0x00, 0x04, 0x00, 0x12, 0x08, 0x25, 0x5c, 0x08, 0x00, 0x08, 0x00, 0x12, 0x08, 0x25, 0x5c, 0x0c, 0x00, 0x0c, 0x00, 0x12, 0x08, 0x25, 0x5c, 0x10, 0x00, 0x10, 0x00, 0x12, 0x08, 0x25, 0x5c,
+	0x14, 0x00, 0x14, 0x00, 0x12, 0x08, 0x25, 0x5c, 0x18, 0x00, 0x18, 0x00, 0x12, 0x08, 0x25, 0x5c, 0x5b, 0x4c, 0x00, 0x49, 0xff, 0x5b, 0x1a, 0x03, 0x08, 0x00, 0x00, 0x00, 0x08, 0x00, 0x4c, 0x00,
+	0x32, 0xff, 0x36, 0x5b, 0x12, 0x00, 0x9e, 0xff, 0x12, 0x00, 0x5e, 0xfc, 0x12, 0x00, 0x18, 0x00, 0xb7, 0x08, 0x01, 0x00, 0x00, 0x00, 0x10, 0x27, 0x00, 0x00, 0x1b, 0x00, 0x01, 0x00, 0x09, 0x00,
+	0xfc, 0xff, 0x01, 0x00, 0x02, 0x5b, 0x1a, 0x03, 0x04, 0x00, 0xf0, 0xff, 0x00, 0x00, 0x4c, 0x00, 0xe0, 0xff, 0x5c, 0x5b, 0x11, 0x14, 0xd6, 0xff, 0x11, 0x04, 0x02, 0x00, 0x30, 0xa0, 0x00, 0x00,
+	0x11, 0x04, 0x02, 0x00, 0x30, 0xe1, 0x00, 0x00, 0x30, 0x41, 0x00, 0x00, 0x11, 0x00, 0x02, 0x00, 0x2b, 0x09, 0x29, 0x00, 0x04, 0x00, 0x01, 0x00, 0x02, 0x00, 0x70, 0x00, 0x01, 0x00, 0x08, 0x00,
+	0x00, 0x00, 0x64, 0x00, 0xff, 0xff, 0x15, 0x07, 0x08, 0x00, 0x0b, 0x5b, 0xb7, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x10, 0x00, 0x15, 0x07, 0x18, 0x00, 0x4c, 0x00, 0xec, 0xfb, 0x0b, 0x5b,
+	0x1b, 0x07, 0x18, 0x00, 0x09, 0x00, 0xf8, 0xff, 0x01, 0x00, 0x4c, 0x00, 0xea, 0xff, 0x5c, 0x5b, 0x1a, 0x07, 0x10, 0x00, 0xec, 0xff, 0x00, 0x00, 0x08, 0x08, 0x4c, 0x00, 0xd0, 0xff, 0x08, 0x5b,
+	0xb7, 0x08, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x10, 0x00, 0x1b, 0x03, 0x04, 0x00, 0x09, 0x00, 0xfc, 0xff, 0x01, 0x00, 0x08, 0x5b, 0x1a, 0x03, 0x0c, 0x00, 0xf0, 0xff, 0x00, 0x00, 0x08, 0x08,
+	0x4c, 0x00, 0xde, 0xff, 0x5c, 0x5b, 0x1a, 0x07, 0x70, 0x00, 0x00, 0x00, 0x22, 0x00, 0x4c, 0x00, 0x98, 0xfb, 0x4c, 0x00, 0x94, 0xfb, 0x36, 0x40, 0x4c, 0x00, 0x9a, 0xfb, 0x36, 0x08, 0x08, 0x08,
+	0x08, 0x40, 0x4c, 0x00, 0x82, 0xff, 0x36, 0x36, 0x4c, 0x00, 0xe8, 0xfb, 0x5c, 0x5b, 0x11, 0x00, 0x1a, 0xfc, 0x12, 0x00, 0x9c, 0xff, 0x12, 0x00, 0xbe, 0xff, 0x12, 0x00, 0xba, 0xff, 0x11, 0x0c,
+	0x08, 0x5c, 0x11, 0x00, 0x02, 0x00, 0x2b, 0x09, 0x29, 0x54, 0x0c, 0x00, 0x01, 0x00, 0x02, 0x00, 0x90, 0x00, 0x01, 0x00, 0x06, 0x00, 0x00, 0x00, 0x6a, 0xfd, 0xff, 0xff, 0x11, 0x00, 0x02, 0x00,
+	0x2b, 0x09, 0x29, 0x00, 0x04, 0x00, 0x01, 0x00, 0x02, 0x00, 0x1c, 0x00, 0x01, 0x00, 0x01, 0x00, 0x00, 0x00, 0x04, 0x00, 0xff, 0xff, 0x16, 0x03, 0x1c, 0x00, 0x4b, 0x5c, 0x46, 0x5c, 0x00, 0x00,
+	0x00, 0x00, 0x11, 0x00, 0xc6, 0xfb, 0x46, 0x5c, 0x04, 0x00, 0x04, 0x00, 0x11, 0x08, 0x22, 0x5c, 0x5b, 0x08, 0x08, 0x4c, 0x00, 0x13, 0xfb, 0x08, 0x5c, 0x5b, 0x11, 0x00, 0x02, 0x00, 0x2b, 0x09,
+	0x29, 0x00, 0x04, 0x00, 0x01, 0x00, 0x02, 0x00, 0x1c, 0x00, 0x01, 0x00, 0x01, 0x00, 0x00, 0x00, 0x2e, 0x00, 0xff, 0xff, 0xb7, 0x08, 0x01, 0x00, 0x00, 0x00, 0x10, 0x27, 0x00, 0x00, 0x1b, 0x03,
+	0x04, 0x00, 0x19, 0x00, 0x14, 0x00, 0x01, 0x00, 0x4b, 0x5c, 0x48, 0x49, 0x04, 0x00, 0x00, 0x00, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x12, 0x08, 0x25, 0x5c, 0x5b, 0x08, 0x5c, 0x5b, 0x1a, 0x03,
+	0x1c, 0x00, 0x00, 0x00, 0x0e, 0x00, 0x08, 0x08, 0x08, 0x08, 0x08, 0x4c, 0x00, 0xc7, 0xff, 0x36, 0x5c, 0x5b, 0x12, 0x00, 0xca, 0xff, 0x11, 0x04, 0x02, 0x00, 0x2b, 0x09, 0x29, 0x54, 0x0c, 0x00,
+	0x01, 0x00, 0x02, 0x00, 0x04, 0x00, 0x01, 0x00, 0x01, 0x00, 0x00, 0x00, 0x52, 0xfd, 0xff, 0xff, 0x11, 0x00, 0x02, 0x00, 0x2b, 0x09, 0x29, 0x00, 0x04, 0x00, 0x01, 0x00, 0x02, 0x00, 0x08, 0x00,
+	0x01, 0x00, 0x01, 0x00, 0x00, 0x00, 0x04, 0x00, 0xff, 0xff, 0x16, 0x03, 0x08, 0x00, 0x4b, 0x5c, 0x46, 0x5c, 0x00, 0x00, 0x00, 0x00, 0x12, 0x08, 0x25, 0x5c, 0x5b, 0x08, 0x08, 0x5b, 0x11, 0x04,
+	0x02, 0x00, 0x2b, 0x09, 0x29, 0x54, 0x0c, 0x00, 0x01, 0x00, 0x02, 0x00, 0x08, 0x00, 0x01, 0x00, 0x02, 0x00, 0x00, 0x00, 0xe2, 0xfd, 0xff, 0xff, 0x00
 }};
 #endif
+
+void SRV_OpnumNotImplemented(handle_t IDL_handle)
+{
+}
+
+ULONG SRV_IDL_DRSCrackNamesNotImplemented(DRS_HANDLE hDrs, DWORD dwInVersion, DRS_MSG_CRACKREQ *pmsgIn, DWORD *pdwOutVersion, DRS_MSG_CRACKREPLY *pmsgOut)
+{
+	return STATUS_NOT_IMPLEMENTED;
+}
+
+ULONG SRV_IDL_DRSDomainControllerInfoNotImplemented(DRS_HANDLE hDrs, DWORD dwInVersion, DRS_MSG_DCINFOREQ *pmsgIn, DWORD *pdwOutVersion, DRS_MSG_DCINFOREPLY *pmsgOut)
+{
+	return STATUS_NOT_IMPLEMENTED;
+}
